Effectful.State.Extra

Description

Compatibility instances for Effectful effects. Particularly useful with the optics functions that operate on Monad[ReaderStateWriter]

Documentation

use' :: forall k a (r :: [Effect]) (is :: IxList) b. (Is k A_Getter, State a :> r) => Optic' k is a b -> Eff r b

locally :: forall s (r :: [Effect]) a. State s :> r => (s -> s) -> Eff r a -> Eff r a

scoped :: forall s (r :: [Effect]) a. State s :> r => Eff r a -> Eff r a

Makes a State effect scoped, so that any changes to the state are reverted after the computation finishes.