Elara.AST.Desugared

Description

This is the second main AST stage, which is very similar to the Expr AST, with a few key differences:

Lambdas only have 1 argument (ones with multiple arguments are desugared into nested lambdas)

Let bindings have no patterns, they are desugared into lambdas

Def and Let declarations are merged into a single entity

Documentation

type DesugaredExpr = Expr 'Desugared

type DesugaredExpr' = Expr' 'Desugared

type DesugaredPattern = Pattern 'Desugared

type DesugaredPattern' = Pattern' 'Desugared

type DesugaredBinaryOperator = BinaryOperator 'Desugared

type DesugaredBinaryOperator' = BinaryOperator' 'Desugared

type DesugaredType = Type 'Desugared

type DesugaredType' = Type' 'Desugared

type DesugaredDeclaration = Declaration 'Desugared

type DesugaredDeclaration' = Declaration' 'Desugared

type DesugaredDeclarationBody = DeclarationBody 'Desugared

type DesugaredDeclarationBody' = DeclarationBody' 'Desugared

type DesugaredTypeDeclaration = TypeDeclaration 'Desugared