Elara.AST.Generic.Types

Documentation

type family Select (s :: ASTSelector) (ast :: a) = (v :: Type)

Used to select a field type for a given AST.

Conventions for usage: If a selection is likely to be one of the principal newtypes (Expr, Pattern, etc), it should not be wrapped in ASTLocate, as this increases friction and creates redundant Located wrappers. This means that implementations should manually wrap in Locate if not using one of the principal newtypes

data Expr' (ast :: a)

Constructors

Int Integer

Float Double

String Text

Char Char

Unit

Var (ASTLocate ast (Select 'ASTVarRef ast))

Constructor (ASTLocate ast (Select 'ConRef ast))

Lambda (ASTLocate ast (Select 'LambdaPattern ast)) (Expr ast)

FunctionCall (Expr ast) (Expr ast)

TypeApplication (Expr ast) (Select 'TypeApplication ast)

If (Expr ast) (Expr ast) (Expr ast)

BinaryOperator !(Select 'ASTBinaryOperator ast)

List !(Select 'List ast)

Match (Expr ast) [(Pattern ast, Expr ast)]

LetIn

Fields

letInParam :: ASTLocate ast (Select 'LetParamName ast)

letInPattern :: Select 'LetPattern ast

letInValue :: Expr ast

letInBody :: Expr ast

Let (ASTLocate ast (Select 'LetParamName ast)) (Select 'LetPattern ast) (Expr ast)

Block (NonEmpty (Expr ast))

Tuple !(Select 'Tuple ast)

InParens !(Select 'InParens ast)

newtype Expr (ast :: a)

Constructors

Expr (ASTLocate ast (Expr' ast), Select ('ASTType 'ForExpr) ast)

data Pattern' (ast :: a)

Constructors

VarPattern (ASTLocate ast (Select 'VarPat ast))

ConstructorPattern (ASTLocate ast (Select 'ConPat ast)) [Pattern ast]

ListPattern !(Select 'ListPattern ast)

TuplePattern !(Select 'TuplePattern ast)

ConsPattern !(Select 'ConsPattern ast)

WildcardPattern

IntegerPattern Integer

FloatPattern Double

StringPattern Text

CharPattern Char

UnitPattern

newtype Pattern (ast :: a)

Constructors

Pattern (ASTLocate ast (Pattern' ast), Select 'PatternType ast)

newtype TypedLambdaParam v (ast :: a)

Wrapper over a tuple of a param name and its type Used mainly to influence Prettyprinting (if we just used a tuple it would get printed as (x,t) rather than x:t)

Constructors

TypedLambdaParam (v, Select 'PatternType ast)

newtype Type (ast :: a)

Constructors

Type (ASTLocate ast (Type' ast), Select 'TypeKind ast)

data Type' (ast :: a)

Constructors

TypeVar (ASTLocate ast (Select 'ASTTypeVar ast))

FunctionType (Type ast) (Type ast)

UnitType

TypeConstructorApplication (Type ast) (Type ast)

UserDefinedType (ASTLocate ast (Select 'UserDefinedType ast))

RecordType (NonEmpty (ASTLocate ast LowerAlphaName, Type ast))

TupleType (NonEmpty (Type ast))

ListType (Type ast)

newtype BinaryOperator (ast :: a)

Constructors

MkBinaryOperator (ASTLocate ast (BinaryOperator' ast))

data BinaryOperator' (ast :: a)

Constructors

SymOp (ASTLocate ast (Select 'SymOp ast))

Infixed (Select 'Infixed ast)

newtype DeclarationBody (ast :: a)

Constructors

DeclarationBody (ASTLocate ast (DeclarationBody' ast))

data DeclarationBody' (ast :: a)

Constructors

Value
let p = e

Fields

_valueName :: ASTLocate ast (Select ('ASTName 'ForValueDecl) ast)

_expression :: Expr ast

_patterns :: Select ('Patterns 'ForValueDecl) ast

_valueType :: Select ('ASTType 'ForValueDecl) ast

_valueAnnotations :: ValueDeclAnnotations ast

ValueTypeDef
def name : type.

Fields

_valueName :: ASTLocate ast (Select ('ASTName 'ForValueDecl) ast)

_valueTypeDef :: !(Select 'ValueTypeDef ast)

_valueTypeDefAnnotations :: ValueDeclAnnotations ast

TypeDeclaration
type name vars = type

Fields

_typeDeclarationName :: ASTLocate ast (Select ('ASTName 'ForType) ast)

typeVars :: [ASTLocate ast (Select 'ASTTypeVar ast)]

typeDeclarationBody :: ASTLocate ast (TypeDeclaration ast)

typeAnnotations :: TypeDeclAnnotations ast

data TypeDeclaration (ast :: a)

Constructors

ADT (NonEmpty (ASTLocate ast (Select 'ConstructorName ast), [Select 'ADTParam ast]))

Alias !(Select 'Alias ast)

newtype Declaration (ast :: a)

Constructors

Declaration (ASTLocate ast (Declaration' ast))

data Declaration' (ast :: a)

Constructors

Declaration'

Fields

moduleName :: ASTLocate ast ModuleName

body :: DeclarationBody ast

newtype ValueDeclAnnotations (ast :: a)

Constructors

ValueDeclAnnotations

Fields

valueAnnotations :: Select ('Annotations 'ForValueDecl) ast

data TypeDeclAnnotations (ast :: a)

Constructors

TypeDeclAnnotations

Fields

kindAnn :: !(Select 'KindAnnotation ast)

typeDeclAnnotations :: !(Select ('Annotations 'ForTypeDecl) ast)

data Annotation (ast :: a)

Constructors

Annotation

Fields

annotationName :: ASTLocate ast (Select 'AnnotationName ast)

annotationArgs :: [AnnotationArg ast]

newtype AnnotationArg (ast :: a)

An annotation argument This is wrapper around an expression with the invariant that it must be a constant value

Constructors

AnnotationArg

Fields

annotationArgValue :: Expr ast

typeOf :: forall {a} (ast :: a). Expr ast -> Select ('ASTType 'ForExpr) ast

freeTypeVars :: forall {a1} {a2} (ast1 :: a1) {ast2 :: a2} {b}. (CleanupLocated (ASTLocate' ast2 (Select 'ASTTypeVar ast2)) ~ Located b, CleanupLocated (ASTLocate' ast1 (Select 'ASTTypeVar ast1)) ~ Located b, CleanupLocated (ASTLocate' ast1 (Type' ast1)) ~ Located (Type' ast2), Ord b, GPlateImpl (K1 R b :: Type -> Type) (Type' ast2), GPlateImpl (K1 R (Select 'TypeKind ast2) :: Type -> Type) (Type' ast2), GPlateImpl (K1 R (CleanupLocated (ASTLocate' ast2 (Select 'UserDefinedType ast2))) :: Type -> Type) (Type' ast2), GPlateImpl (K1 R (CleanupLocated (ASTLocate' ast2 (Type' ast2))) :: Type -> Type) (Type' ast2), GPlateImpl (K1 R (CleanupLocated (ASTLocate' ast2 LowerAlphaName)) :: Type -> Type) (Type' ast2)) => Type ast1 -> [CleanupLocated (ASTLocate' ast1 (Select 'ASTTypeVar ast1))]

patternTypeOf :: forall {a} (ast :: a). Pattern ast -> Select 'PatternType ast

class RUnlocate (ast1 :: ast) where

The idea of this class is to provide a way to work under both located and unlocated ASTs We define functions which are polymorphic over the location strategy used in the AST For example, rUnlocate provides a way to extract the unlocated value from an AST node, regardless of whether the AST is located or unlocated

Minimal complete definition

rUnlocated, fmapUnlocated, traverseUnlocated

Methods

rUnlocate :: CleanupLocated (Located a) ~ Located a => ASTLocate ast1 a -> a

rUnlocated :: CleanupLocated (Located a) ~ Located a => Getter (ASTLocate ast1 a) a

fmapUnlocated :: (CleanupLocated (Located a) ~ Located a, CleanupLocated (Located b) ~ Located b) => (a -> b) -> ASTLocate ast1 a -> ASTLocate ast1 b

traverseUnlocated :: (CleanupLocated (Located a) ~ Located a, CleanupLocated (Located b) ~ Located b) => Traversal (ASTLocate ast1 a) (ASTLocate ast1 b) a b

type ASTLocate (ast :: a) a1 = CleanupLocated (ASTLocate' ast a1)

type family ASTLocate' (ast :: a) :: Type -> Type

type FullASTQual (ast :: a) a1 = ASTLocate ast (ASTQual ast a1)

type family ASTQual (ast :: a) :: Type -> Type

type family CleanupLocated g where ...

Unwraps a single layer of Unlocated from a type.

Equations

CleanupLocated (Unlocated a) = a

CleanupLocated (Located (Located a)) = CleanupLocated a

CleanupLocated (Located (Expr a2)) = CleanupLocated (Expr a2)

CleanupLocated (Located (Pattern a2)) = CleanupLocated (Pattern a2)

CleanupLocated (Located (BinaryOperator a2)) = CleanupLocated (BinaryOperator a2)

CleanupLocated (Located (Declaration a2)) = CleanupLocated (Declaration a2)

CleanupLocated (Located (DeclarationBody a2)) = CleanupLocated (DeclarationBody a2)

CleanupLocated (Located (Type a2)) = CleanupLocated (Type a2)

CleanupLocated a = a

newtype Unlocated a

Constructors

Unlocated a

coerceType :: forall {a1} {a2} (ast1 :: a1) (ast2 :: a2) {f :: Type -> Type}. (Select 'TypeKind ast1 ~ Select 'TypeKind ast2, CleanupLocated (ASTLocate' ast2 (Type' ast2)) ~ f (Type' ast2), CleanupLocated (ASTLocate' ast1 (Select 'ASTTypeVar ast1)) ~ CleanupLocated (ASTLocate' ast2 (Select 'ASTTypeVar ast2)), CleanupLocated (ASTLocate' ast1 (Select 'UserDefinedType ast1)) ~ CleanupLocated (ASTLocate' ast2 (Select 'UserDefinedType ast2)), CleanupLocated (ASTLocate' ast1 (Type' ast1)) ~ f (Type' ast1), CleanupLocated (ASTLocate' ast1 LowerAlphaName) ~ CleanupLocated (ASTLocate' ast2 LowerAlphaName), Functor f) => Type ast1 -> Type ast2

coerceType' :: forall {a1} {a2} (ast1 :: a1) (ast2 :: a2) {f :: Type -> Type}. (Select 'TypeKind ast1 ~ Select 'TypeKind ast2, CleanupLocated (ASTLocate' ast2 (Type' ast2)) ~ f (Type' ast2), CleanupLocated (ASTLocate' ast1 (Select 'ASTTypeVar ast1)) ~ CleanupLocated (ASTLocate' ast2 (Select 'ASTTypeVar ast2)), CleanupLocated (ASTLocate' ast1 (Select 'UserDefinedType ast1)) ~ CleanupLocated (ASTLocate' ast2 (Select 'UserDefinedType ast2)), CleanupLocated (ASTLocate' ast1 (Type' ast1)) ~ f (Type' ast1), CleanupLocated (ASTLocate' ast1 LowerAlphaName) ~ CleanupLocated (ASTLocate' ast2 LowerAlphaName), Functor f) => Type' ast1 -> Type' ast2

coerceTypeDeclaration :: forall {a1} {a2} {a3} {a4} {a5} {a6} (ast4 :: a1) (ast5 :: a2) {ast6 :: a3} {ast7 :: a4} {ast8 :: a5} {ast9 :: a6} {f1 :: Type -> Type} {f2 :: Type -> Type}. (Select 'TypeKind ast6 ~ Select 'TypeKind ast7, Select 'Alias ast4 ~ Type ast6, Select 'ADTParam ast4 ~ Type ast8, Select 'TypeKind ast8 ~ Select 'TypeKind ast9, Select 'ADTParam ast5 ~ Type ast9, Select 'Alias ast5 ~ Type ast7, CleanupLocated (ASTLocate' ast8 (Select 'UserDefinedType ast8)) ~ CleanupLocated (ASTLocate' ast9 (Select 'UserDefinedType ast9)), CleanupLocated (ASTLocate' ast8 (Select 'ASTTypeVar ast8)) ~ CleanupLocated (ASTLocate' ast9 (Select 'ASTTypeVar ast9)), CleanupLocated (ASTLocate' ast8 (Type' ast8)) ~ f1 (Type' ast8), CleanupLocated (ASTLocate' ast8 LowerAlphaName) ~ CleanupLocated (ASTLocate' ast9 LowerAlphaName), CleanupLocated (ASTLocate' ast9 (Type' ast9)) ~ f1 (Type' ast9), CleanupLocated (ASTLocate' ast7 (Type' ast7)) ~ f2 (Type' ast7), CleanupLocated (ASTLocate' ast6 (Select 'ASTTypeVar ast6)) ~ CleanupLocated (ASTLocate' ast7 (Select 'ASTTypeVar ast7)), CleanupLocated (ASTLocate' ast6 (Select 'UserDefinedType ast6)) ~ CleanupLocated (ASTLocate' ast7 (Select 'UserDefinedType ast7)), CleanupLocated (ASTLocate' ast6 (Type' ast6)) ~ f2 (Type' ast6), CleanupLocated (ASTLocate' ast6 LowerAlphaName) ~ CleanupLocated (ASTLocate' ast7 LowerAlphaName), CleanupLocated (ASTLocate' ast4 (Select 'ConstructorName ast4)) ~ CleanupLocated (ASTLocate' ast5 (Select 'ConstructorName ast5)), Functor f2, Functor f1) => TypeDeclaration ast4 -> TypeDeclaration ast5

pattern Expr' :: RUnlocate ast => Expr' ast -> Expr ast

exprLocation :: forall {a} (ast :: a). ASTLocate' ast ~ Located => Lens' (Expr ast) SourceRegion

patternLocation :: forall {a} (ast :: a). ASTLocate' ast ~ Located => Lens' (Pattern ast) SourceRegion

coerceTypeDeclAnnotations :: forall {a1} {a2} (ast1 :: a1) (ast2 :: a2). (Select ('Annotations 'ForTypeDecl) ast1 ~ [Annotation ast1], Select ('Annotations 'ForTypeDecl) ast2 ~ [Annotation ast2], Select 'KindAnnotation ast1 ~ Select 'KindAnnotation ast2) => TypeDeclAnnotations ast1 -> TypeDeclAnnotations ast2

traverseValueDeclAnnotations :: forall {a1} {a2} (ast1 :: a1) (t :: Type -> Type) (ast2 :: a2) f. (Select ('Annotations 'ForValueDecl) ast1 ~ t (Annotation ast1), Select ('Annotations 'ForValueDecl) ast2 ~ t (Annotation ast2), Traversable t, Applicative f) => (Annotation ast1 -> f (Annotation ast2)) -> ValueDeclAnnotations ast1 -> f (ValueDeclAnnotations ast2)

traverseTypeDeclAnnotations :: forall {a1} {a2} (ast1 :: a1) (t :: Type -> Type) (ast2 :: a2) f. (Select ('Annotations 'ForTypeDecl) ast1 ~ t (Annotation ast1), Select ('Annotations 'ForTypeDecl) ast2 ~ t (Annotation ast2), Traversable t, Select 'KindAnnotation ast1 ~ Select 'KindAnnotation ast2, Applicative f) => (Annotation ast1 -> f (Annotation ast2)) -> TypeDeclAnnotations ast1 -> f (TypeDeclAnnotations ast2)

declarationBody'Name :: forall {a} (ast :: a). (CleanupLocated (Located (Select ('ASTName 'ForType) ast)) ~ Located (Select ('ASTName 'ForType) ast), CleanupLocated (Located (Select ('ASTName 'ForValueDecl) ast)) ~ Located (Select ('ASTName 'ForValueDecl) ast), RUnlocate ast, ToName (Select ('ASTName 'ForValueDecl) ast), ToName (Select ('ASTName 'ForType) ast)) => Getter (DeclarationBody' ast) (ASTLocate ast Name)

declarationBodyName :: forall {a} (ast :: a). (CleanupLocated (Located (Select ('ASTName 'ForValueDecl) ast)) ~ Located (Select ('ASTName 'ForValueDecl) ast), CleanupLocated (Located (Select ('ASTName 'ForType) ast)) ~ Located (Select ('ASTName 'ForType) ast), RUnlocate ast, ToName (Select ('ASTName 'ForValueDecl) ast), ToName (Select ('ASTName 'ForType) ast)) => Getter (DeclarationBody ast) (CleanupLocated (ASTLocate' ast Name))

declaration'Name :: forall {a} (ast :: a). (CleanupLocated (Located (Select ('ASTName 'ForValueDecl) ast)) ~ Located (Select ('ASTName 'ForValueDecl) ast), CleanupLocated (Located (Select ('ASTName 'ForType) ast)) ~ Located (Select ('ASTName 'ForType) ast), RUnlocate ast, ToName (Select ('ASTName 'ForValueDecl) ast), ToName (Select ('ASTName 'ForType) ast)) => Getter (Declaration' ast) (CleanupLocated (ASTLocate' ast Name))

declarationName :: forall {a} (ast :: a). (CleanupLocated (Located (Select ('ASTName 'ForType) ast)) ~ Located (Select ('ASTName 'ForType) ast), CleanupLocated (Located (Select ('ASTName 'ForValueDecl) ast)) ~ Located (Select ('ASTName 'ForValueDecl) ast), RUnlocate ast, ToName (Select ('ASTName 'ForValueDecl) ast), ToName (Select ('ASTName 'ForType) ast)) => Getter (Declaration ast) (CleanupLocated (ASTLocate' ast Name))