Elara.AST.Generic.Utils

Documentation

class ToMaybe i o where

When fields may be optional, we need a way of representing that generally. This class does that. In short, it converts a type to a Maybe. If the type is already a Maybe, it is left alone. If it is not, it is wrapped in a Just. If it is NoFieldValue, it is converted to Nothing.

Methods

toMaybe :: i -> o

class ToList i o where

Like ToMaybe but for lists. Useful when fields may be lists or single values.

Methods

fieldToList :: i -> o

class ApplyAsFunctorish i o a b where

Sometimes fields are wrapped in functors eg lists, we need a way of transcending them. This class does that. For example, let's say we have `cleanPattern :: Pattern ast1 -> Pattern ast2`, and `x :: Select ast1 Pattern`. x could be `Pattern ast1`, `[Pattern ast1]`, `Maybe (Pattern ast1)`, or something else entirely. cleanPattern will only work on the first of these, so we need a way of lifting it to the others. Obviously, this sounds like a Functor, but `Pattern ast1` has the wrong kind for a functor. This class half-solves that problem by allowing us to lift a function to a functorish type, whether it is a functor or not.

Methods

applyAsFunctorish :: (a -> b) -> i -> o

class DataConAs a b where

A class for converting between a data constructor and a type, useful for types that may be DataConCantHappen

Methods

dataConAs :: a -> b

asDataCon :: b -> a

type family UnwrapMaybe a = (k :: Type) where ...

Unwraps 1 level of Maybe from a type. Useful when a type family returns Maybe

Equations

UnwrapMaybe (Maybe a) = a

UnwrapMaybe a = a

type family UnwrapList a = (k :: Type) where ...

Unwraps 1 level of '[]' from a type. Useful when a type family returns []

Equations

UnwrapList [a] = a

UnwrapList a = a