Elara.AST.Introspection

Description

Defines useful classes and functions for introspecting ASTs

Documentation

class IntrospectableAnnotations (ast :: a) where

Value-level proof that an AST's annotations can be coerced to a specific type over all selectors

Methods

getAnnotations :: forall (sel :: ForSelector). Select ('Annotations sel) ast -> [Annotation ast]

Get all annotations for a given selector in the AST