Elara.AST.Location

Documentation

data AstNode

The different kinds of nodes that should have incompatible locations to prevent accidentally using the wrong location for a node. the actual definitions of what is in each category is a bit vague

Constructors

ModuleNode
a location of the _whole_ module

DeclNode
a location of an entire declaration

ExprNode
a location of an expression

PatternNode
a location of a pattern

TypeNode
a location of a type

VarNode
a location of a variable

data family NodeLoc (n :: AstNode) loc

class UnwrapNodeLoc (n :: AstNode) loc a where

Methods

unwrapLoc :: NodeLoc n loc -> a

type family Join (a :: AstNode) (b :: AstNode) :: AstNode where ...

the "join" of two nodes, i.e. the smallest node that subsumes both of them. note that this is not the same as a "container" relationship. for example, an ExprNode can contain a PatternNode, since a pattern can appear inside an expression, but the join of ExprNode and PatternNode is ExprNode, since ExprNode is the only node that could contain both at once. a should be the smaller node and b the larger node. by default, this relation is symmetric, so should be overridden for asymmetric cases.

Equations

Join n n = n

Join 'ExprNode 'PatternNode = 'ExprNode

Join 'VarNode 'ExprNode = 'DeclNode

Join 'VarNode 'TypeNode = 'TypeNode

Join 'DeclNode n = 'DeclNode

Join n 'DeclNode = 'DeclNode

Join 'ModuleNode _1 = 'ModuleNode

Join a b = Join b a

type Subsumes (narrower :: AstNode) (wider :: AstNode) = Join narrower wider ~ wider

A node a subsumes a node b if the join of a and b is a

class Subsumes a b => Widen (a :: AstNode) (b :: AstNode) where

Widen a node to a wider node at the data level

Methods

widen :: NodeLoc a loc -> NodeLoc b loc

class LocSemigroup (a :: AstNode) (b :: AstNode) loc where

Methods

(<.>) :: NodeLoc a loc -> NodeLoc b loc -> NodeLoc (Join a b) loc infixr 6

class HasLocation a loc | a -> loc where

Methods

getLocation :: a -> loc

class WrapNode (n :: AstNode) where

class for wrapping a raw location in a node-specific location wrapper

Methods

wrap :: loc -> NodeLoc n loc

data TaggedLocate (n :: AstNode) loc a

a version of Located that is tagged with the specific node type it is located for

Constructors

TaggedLocate !(NodeLoc n loc) a

tag :: forall (n :: AstNode) -> forall a loc. (HasLocation a loc, WrapNode n) => a -> NodeLoc n loc

Tag a value that HasLocation with a specific located node type

tagLocated :: forall (n :: AstNode) a. WrapNode n => Located a -> TaggedLocate n SourceRegion a

Convert a generic Located value into a tagged TaggedLocate

spanAs :: forall (n :: AstNode) a loc. (HasLocation a loc, WrapNode n, Monoid loc) => [a] -> NodeLoc n loc

retag :: forall (newTag :: AstNode) (oldTag :: AstNode) loc a. (UnwrapNodeLoc oldTag loc loc, WrapNode newTag) => TaggedLocate oldTag loc a -> TaggedLocate newTag loc a

stripTag :: forall (n :: AstNode) loc a. UnwrapNodeLoc n loc SourceRegion => TaggedLocate n loc a -> Located a