Elara.AST.Name

Documentation

newtype ModuleName

Constructors

ModuleName (NonEmpty Text)

appendModule :: ModuleName -> Text -> ModuleName

data VarName

A valid Variable name. This includes anything that could appear in `let [name] = ...` In other words, a normal alphanumeric name, or a parenthesis wrapped operator name

Constructors

NormalVarName LowerAlphaName
A normal alphanumeric name

OperatorVarName OpName
An operator var name. Note that while in the source code, the name must be surrounded in parentheses, this is not the case in the AST!

newtype LowerAlphaName

A lowercase alphanumeric name. Could be used for variables or type variables Since type variables can't be operators though, we don't use VarName for them

Constructors

LowerAlphaName Text

newtype TypeName

Constructors

TypeName Text

newtype OpName

Constructors

OpName Text

data VarOrConName

Constructors

VarName LowerAlphaName

ConName TypeName

data Name

Constructors

NVarName VarName

NTypeName TypeName

type UniqueName = Unique Name

class NameLike name where

Minimal complete definition

nameText, fullNameText

Methods

nameText :: name -> Text

Get the name as a Text. This will not include qualification, if present

fullNameText :: name -> Text

Get the full name, including qualification, if present

moduleName :: name -> Maybe ModuleName

class ToName name where

Methods

toName :: name -> Name

data MaybeQualified name

Constructors

MaybeQualified

Fields

_maybeQualifiedName :: name

_maybeQualifiedQualifier :: Maybe ModuleName

data Qualified name

Constructors

Qualified

Fields

_qualifiedName :: name

qualifier :: ModuleName

unqualified :: Lens' (Qualified name) name

newtype Unqualified name

Constructors

Unqualified

Fields

_unqualifiedName :: name

class HasName a n | a -> n where

Methods

name :: Lens' a n

class ContainsName a n | a -> n where

Methods

containedName :: Getter a n