Elara.AST.Phases.Frontend.Pretty

Description

Round-trip pretty-printing for Frontend AST. Pretty instances here produce fully-parenthesized output suitable for property tests that check parse . pretty == id. These are necessarily orphan instances (Frontend defined in parent module, Pretty class defined in Elara.Data.Pretty).

Documentation

prettyExprRoundTrip :: PrettyPhaseLoc Frontend loc => Expr loc Frontend -> Doc AnsiStyle

prettyPatternRoundTrip :: PrettyPhaseLoc Frontend loc => Pattern loc Frontend -> Doc AnsiStyle

Orphan instances

PrettyPhaseLoc Frontend loc => Pretty (Expr loc Frontend)

Instance details

Methods

pretty :: Expr loc Frontend -> Doc AnsiStyle

PrettyPhaseLoc Frontend loc => Pretty (Pattern loc Frontend)

Instance details

Methods

pretty :: Pattern loc Frontend -> Doc AnsiStyle

PrettyPhaseLoc Frontend loc => Pretty (Type loc Frontend)

Instance details

Methods

pretty :: Type loc Frontend -> Doc AnsiStyle