Elara.AST.Pretty

Description

Common module for all the different AST pretty impls. This includes functions for pretty printing in both context-free and context-sensitive forms

Documentation

none :: [()]

nothing :: Maybe ()

A Nothing value for Maybe Docs Useful for avoiding type annotations

prettyStringExpr :: Text -> Doc AnsiStyle

prettyCharExpr :: Char -> Doc AnsiStyle

prettyLambdaExpr :: (?contextFree :: Bool, Pretty a, Pretty b) => [a] -> b -> Doc AnsiStyle

prettyFunctionCall :: (?contextFree :: Bool, Pretty a, Pretty b) => a -> b -> Doc AnsiStyle

prettyFunctionCallExpr :: forall {a} (ast :: a). (?contextFree :: Bool, RUnlocate ast, Pretty (Expr ast)) => Expr ast -> Expr ast -> Bool -> Doc AnsiStyle

prettyIfExpr :: (Pretty a, Pretty b, Pretty c) => a -> b -> c -> Doc AnsiStyle

prettyBinaryOperatorExpr :: forall {a} b (ast :: a). (Pretty b, Pretty (Expr ast), RUnlocate ast) => Expr ast -> b -> Expr ast -> Doc AnsiStyle

prettyTupleExpr :: Pretty a => NonEmpty a -> Doc AnsiStyle

prettyMatchExpr :: (Pretty a1, Pretty a2, Foldable t, ?contextFree :: Bool) => a1 -> t a2 -> Doc AnsiStyle

prettyMatchBranch :: (Pretty a1, Pretty a2) => (a1, a2) -> Doc AnsiStyle

prettyLetInExpr :: forall {a} a1 a2 (ast :: a). (Pretty a1, Pretty a2, ?contextFree :: Bool, RUnlocate ast, Pretty (Expr ast)) => a1 -> [a2] -> Expr ast -> Expr ast -> Doc AnsiStyle

shouldBrace :: forall astK (ast :: astK). RUnlocate ast => Expr ast -> Bool

shouldParen :: forall astK (ast :: astK). RUnlocate ast => Expr ast -> Bool

prettyLetExpr :: forall {a} a1 a2 (ast :: a). (Pretty a1, Pretty a2, RUnlocate ast, ?contextFree :: Bool, Pretty (Expr ast)) => a1 -> [a2] -> Expr ast -> Doc AnsiStyle

prettyBlockExpr :: (Pretty a, Foldable t, ?contextFree :: Bool) => t a -> Doc AnsiStyle

encloseSep' :: Bool -> Doc AnsiStyle -> Doc AnsiStyle -> Doc AnsiStyle -> [Doc AnsiStyle] -> Doc AnsiStyle

prettyConstructorPattern :: (Pretty a1, Pretty a2) => a1 -> [a2] -> Doc AnsiStyle

prettyList :: (Pretty a, ?contextFree :: Bool) => [a] -> Doc AnsiStyle

prettyConsPattern :: (Pretty a1, Pretty a2) => a1 -> a2 -> Doc AnsiStyle

prettyValueDeclaration :: forall {a} (ast :: a) a1 a2 a3. (Pretty a1, Pretty a2, Pretty a3, Pretty (Select ('Annotations 'ForValueDecl) ast)) => a1 -> a2 -> Maybe a3 -> ValueDeclAnnotations ast -> Doc AnsiStyle

prettyValueDeclAnnotations :: forall {a} (ast :: a). Pretty (Select ('Annotations 'ForValueDecl) ast) => ValueDeclAnnotations ast -> Doc AnsiStyle

prettyValueTypeDef :: forall {a} (ast :: a) a1 a2. (Pretty a1, Pretty a2, Pretty (Select ('Annotations 'ForValueDecl) ast)) => a1 -> a2 -> Maybe (ValueDeclAnnotations ast) -> Doc AnsiStyle

prettyTypeDeclaration :: forall {a} a1 a2 a3 (ast :: a). (Pretty a1, Pretty a2, Pretty a3) => a1 -> [a2] -> a3 -> TypeDeclAnnotations ast -> Doc AnsiStyle