Elara.AST.Region

Documentation

generatedFileName :: String

data RealPosition

Constructors

Position

Fields

_line :: !Int

_column :: !Int

data Position

Constructors

RealPosition !RealPosition

GeneratedPosition

data RealSourceRegion

Constructors

UnsafeMkSourceRegion

Fields

_sourceFile :: !(Maybe FilePath)

_startPos :: !RealPosition

_endPos :: !RealPosition

data SourceRegion

Constructors

RealSourceRegion !RealSourceRegion

GeneratedRegion !FilePath

_RealPosition :: Prism' Position RealPosition

_GeneratedPosition :: Prism' Position ()

column :: Lens' RealPosition Int

line :: Lens' RealPosition Int

_RealSourceRegion :: Prism' SourceRegion RealSourceRegion

_GeneratedRegion :: Prism' SourceRegion FilePath

endPos :: Lens' RealSourceRegion RealPosition

sourceFile :: Lens' RealSourceRegion (Maybe FilePath)

startPos :: Lens' RealSourceRegion RealPosition

class HasPath a where

Methods

path :: Lens' a (Maybe FilePath)

mkSourceRegion :: SourcePos -> SourcePos -> RealSourceRegion

mkSourceRegionIn :: HasCallStack => Maybe FilePath -> RealPosition -> RealPosition -> RealSourceRegion

spToPosition :: SourcePos -> RealPosition

positionToSp :: FilePath -> RealPosition -> SourcePos

generatedSourcePos :: Maybe FilePath -> SourcePos

generatedSourceRegion :: Maybe FilePath -> SourceRegion

generatedLocated :: Maybe FilePath -> a -> Located a

sourceRegionToDiagnosePosition :: SourceRegion -> Position

positionToDiagnosePosition :: FilePath -> RealPosition -> Position

data Located a

Constructors

Located SourceRegion a

class HasSourceRegion a where

Methods

sourceRegion :: Lens' a SourceRegion

_Located :: Iso (Located a1) (Located a2) (SourceRegion, a1) (SourceRegion, a2)

type family Unlocate g where ...

Equations

Unlocate (Located a) = a

Unlocate a = a

newtype IgnoreLocation a

Newtype wrapper for Located that ignores the location information for its instances

Constructors

IgnoreLocation (Located a)

_IgnoreLocation :: Iso (IgnoreLocation a1) (IgnoreLocation a2) (Located a1) (Located a2)

generatedSourceRegionFrom :: Located a -> SourceRegion

unlocated :: Lens (Located a) (Located b) a b

withLocationOf :: HasSourceRegion b => a -> b -> Located a

merge :: (HasSourceRegion a, HasSourceRegion b) => (a -> b -> c) -> a -> b -> Located c

enclosingRegion :: HasCallStack => RealSourceRegion -> RealSourceRegion -> RealSourceRegion

Get the region that contains both of the given regions. This function will throw an error if the regions are in different files.

enclosingRegion' :: HasCallStack => SourceRegion -> SourceRegion -> SourceRegion

Get the region that contains both of the given regions. This function will throw an error if the regions are in different files. If either of the given SourceRegions is a GeneratedRegion, then the result will be a GeneratedRegion.

spanningRegion :: NonEmpty RealSourceRegion -> RealSourceRegion

spanningRegion' :: NonEmpty SourceRegion -> SourceRegion

Get the region that contains all of the given regions. This function will throw an error if the regions are in different files. If all the given SourceRegions are GeneratedRegions, then the result will be a GeneratedRegion. Otherwise, the GeneratedRegions will be ignored.