Elara.AST.Select

Documentation

data LocatedAST

Main type family used to select types for different AST stages. This type in particular represents the main different primary AST stages and is used as a parameter to Select.

Constructors

Frontend
The AST as produced by the parser. Main types in Elara.AST.Frontend.

Desugared
The AST after basic desugaring. Main types in Elara.AST.Desugared.

Renamed
The AST after name resolution. Main types in Elara.AST.Renamed.

Shunted
The AST after shunting (fixing operator precedence). Main types in Elara.AST.Shunted.

MidKinded
The AST while kind inference is being performed. Done as a separate stage for better type safety. Main types in Elara.AST.Kinded.

Kinded
The AST after kind inference, i.e. all types have their kinds known. Main types in Elara.AST.Kinded.

Typed
The AST after full type inference. Main types in Elara.AST.Typed.

data UnlocatedAST

Unlocated counterpart to LocatedAST. Somewhat underutilised at the moment, but intended for testing.

Constructors

UnlocatedFrontend

UnlocatedDesugared

UnlocatedRenamed

UnlocatedShunted

UnlocatedTyped

type family LocatedToUnlocated (ast :: LocatedAST) :: UnlocatedAST where ...

Type family to convert between located and unlocated AST selectors

Equations

LocatedToUnlocated 'Frontend = 'UnlocatedFrontend

LocatedToUnlocated 'Desugared = 'UnlocatedDesugared

LocatedToUnlocated 'Renamed = 'UnlocatedRenamed

LocatedToUnlocated 'Shunted = 'UnlocatedShunted

LocatedToUnlocated 'Typed = 'UnlocatedTyped

type family UnlocatedToLocated (ast :: UnlocatedAST) :: LocatedAST where ...

Type family to convert between unlocated and located AST selectors

Equations

UnlocatedToLocated 'UnlocatedFrontend = 'Frontend

UnlocatedToLocated 'UnlocatedDesugared = 'Desugared

UnlocatedToLocated 'UnlocatedRenamed = 'Renamed

UnlocatedToLocated 'UnlocatedShunted = 'Shunted

UnlocatedToLocated 'UnlocatedTyped = 'Typed

data ASTSelector

Elements of the AST that are parametric by their AST stage This type acts as a selector for the Select type family for these elements

Constructors

ASTVarRef
The type used for references to variables

ConRef
References to type / data constructors

LambdaPattern
Patterns that can be bound in a lambda expression (eg [a] (b, c) -> ...)

TypeApplication
The value stored in type application nodes (usually a Type)

PatternType
The type used to represent an explicit type declaration for patterns, e.g. [Int] in f ([y] :: [Int])

ASTBinaryOperator
Value of binary operator nodes. Replaced with NoFieldValue after shunting

List
Value of list expressions

LetParamName
Type used for the name of a value bound by a let expression

LetPattern
Patterns that can be bound in a let expression

ConstructorName
Type used for the name of a constructor within its declaration only. Often differs to ConRef in what types of qualification it allows

ADTParam
Type of parameters for a data constructor in its declaration, e.g.the second a in type Box a = Box a

Alias
Type used for the RHS of a type alias declaration

Annotations ForSelector
Type of annotations for a selected AST node type

KindAnnotation
Type for Kind annotations upon a type declaration only

ASTTypeVar
Type of type variables used in any way

TypeKind
Type for the actual kind of a type stored within the Type, i.e. at use sites. Will probably be merged with KindAnnotation as they aren't clearly distinguishable

UserDefinedType
Type for a reference to a user defined type, i.e. a named type that is not a primitive. Usually similar/identical to ConRef

ASTName ForSelector
Type of a name of a specific element

Patterns ForSelector
Type of pattern(s) that can be bound by a specific element

ASTType ForSelector
Type of types of specific elements

VarPat
Type used for the value of variable patterns, i.e. a variable name

ConPat
Type used for the value of constructor patterns, i.e. a constructor name

AnnotationName
Type used for the name of an annotation

ValueTypeDef
Type used for the type part in a def name : type statement

Tuple
Type of a tuple expression

InParens
Type of an expression wrapped in parenthesis

ListPattern
A list literal pattern

TuplePattern
A tuple literal pattern

ConsPattern
A cons literal pattern

SymOp
An operator's name that is symbolic, i.e. made of symbols

Infixed
The name of the "operator" in an infix (e.g. a f b) expression

TupleType
The type of a tuple type expression

data ForSelector

Selector to indicate which part of an AST the ASTSelector is being used for. For example, the ASTName selector can be used for type names, value names, or declaration names, which may have different types at different AST stages. A motivating example is that type names may be qualified at use sites, but not at declaration sites.

Constructors

ForType
For a use of a type name

ForTypeDecl
For the declaration of a type name

ForValueDecl
For the declaration of a value name

ForExpr
For a use of a value name