Elara.AST.Shunted

Description

Shunted AST Type This is very similar to Expr except:

Operators are re-shunted to match their defined precedence and associativity

This means there's no need for an InParens token anymore so that's also gone :D

The confusing VarName/OpName bs is also gone. Binary operator invocations are replaced with prefix function calls. This always uses VarName

Documentation

type ShuntedExpr = Expr 'Shunted

type ShuntedExpr' = Expr' 'Shunted

type ShuntedPattern = Pattern 'Shunted

type ShuntedPattern' = Pattern' 'Shunted

type ShuntedBinaryOperator = BinaryOperator 'Shunted

type ShuntedBinaryOperator' = BinaryOperator' 'Shunted

type ShuntedType = Type 'Shunted

type ShuntedType' = Type' 'Shunted

type ShuntedDeclaration = Declaration 'Shunted

type ShuntedDeclaration' = Declaration' 'Shunted

type ShuntedDeclarationBody = DeclarationBody 'Shunted

type ShuntedDeclarationBody' = DeclarationBody' 'Shunted

type ShuntedTypeDeclaration = TypeDeclaration 'Shunted