Elara.Core

Documentation

data TypeVariable

Constructors

TypeVariable

Fields

tvName :: UniqueTyVar

tvKind :: ElaraKind

data Var

Constructors

TyVar TypeVariable

Id (UnlocatedVarRef Text) Type (Maybe DataCon)

data Expr b

Constructors

Var b

Lit Literal

App (Expr b) (Expr b)

TyApp (Expr b) Type

Lam b (Expr b)

TyLam Type (Expr b)

Let (Bind b) (Expr b)

Match (Expr b) (Maybe b) [Alt b]

type CoreExpr = Expr Var

type CoreAlt = Alt Var

type CoreBind = Bind Var

type Bind b = Bind b Expr

type Alt b = (AltCon, [b], Expr b)

data AltCon

Constructors

DataAlt DataCon

LitAlt Literal

DEFAULT

data DataCon

A data constructor.

Constructors

DataCon

Fields

name :: Qualified Text

The name of the data constructor

dataConType :: Type

The type of the data constructor, i.e. `type Foo a = Bar a` would have a data constructor with type `a -> Foo a`

dataConDataType :: TyCon

The type of the data type the data constructor belongs to, i.e. `type Foo a = Bar a` would have a DataCon for `Bar a` with dataConDataType = Foo a. This should be identical to functionTypeResult . dataConType

data Type

Constructors

TyVarTy TypeVariable
A type variable, a

FuncTy Type Type
A function type, t -> t'

AppTy Type Type
An application of a type constructor, TC t

ConTy TyCon
A type constructor, TC

ForAllTy !TypeVariable !Type
A forall quantified type, forall a. T

data TyCon

Information about a type constructor

Constructors

TyCon

Fields

(Qualified Text)

The name of the type constructor

TyConDetails

The details of the type constructor

data TyConDetails

The details of a type constructor, mainly about its definition

Constructors

TyADT
An ADT

Fields

[Qualified Text]

The ids of its DataCons (constructors)

TyAlias Type

Prim

typeArity :: Type -> Int

The arity of a function type

functionTypeArgs :: Type -> [Type]

conTyArgs :: Type -> [Type]

functionTypeResult :: Type -> Type

instantiate :: Type -> Type -> Type

substTypeVar :: TypeVariable -> Type -> Type -> Type

data Literal

Constructors

Int !Integer

String !Text

Char !Char

Double !Double

Unit