Elara.Core.Analysis

Documentation

estimateArity :: CoreExpr -> Int

declaredLambdaArity :: CoreExpr -> Int

findTyCon :: Type -> Maybe TyCon

guesstimateExprType :: (HasCallStack, Pretty Type, Pretty (Expr Var)) => TraceableFn "guesstimateExprType" CoreExpr Type

overForAll :: Type -> (Type -> Type) -> Type

Applies a function over the monotype of a potential forall expression For example, given a type `forall a. a -> a`, overForAll would apply the function to `a -> a` Or for forall a b c. a -> b -> c, it would apply the function to a -> b -> c

varType :: Var -> Type

literalType :: Literal -> Type

class FreeCoreVars (ast :: Type -> Type) where

Methods

freeCoreVars :: Ord b => ast b -> Set b

freeCoreVarsBind :: forall (ast :: Type -> Type) a. (FreeCoreVars ast, Ord a) => Bind a ast -> Set a

freeTypeVars :: Type -> Set TypeVariable