Elara.Core.LiftClosures.Util

Documentation

varTermType :: Var -> Maybe Type

Get the type from a Var, or Nothing for TyVar

isGlobal :: Var -> Bool

Check if a variable is global

isLocal :: Var -> Bool

Check if a variable is local

filterLocals :: Set Var -> Set Var

Filter to only local variables

computeCaptures :: Var -> Set Var -> [Var]

Compute captured variables (free locals minus the binding itself)

isLam :: AExpr b -> Bool

Check if an atom is a lambda

peelLam :: CExpr Var -> ([Var], Expr Var)

Peel arguments from nested lambdas

mkNestedLam :: [Var] -> Expr Var -> Expr Var

Build a nested lambda from a list of arguments and a body

liftedType :: Traversable f => f Var -> Type -> Maybe Type

Create the lifted type for a function given its captures and original type. That is, if the original type is `a -> b` and the captures are of types `c1, c2`, the lifted type will be `c1 -> c2 -> a -> b`.

closedVariables :: FreeCoreVars ast => ast Var -> Set Var

Get the closed local variables in an expression