Elara.Core.ToANF

Documentation

type ToANF (r :: [Effect]) = (UniqueGen :> r, StructuredDebug :> r, Pretty (Expr Var), Pretty Var, Pretty (AExpr Var), Pretty (CExpr Var), Pretty (Expr Var), Pretty Type)

Convert a Core expression to ANF For example: let main = print (f (g (h 1))) becomes: let main = let v1 = h 1 in let v2 = g v1 in let v3 = f v2 in print v3

toANF :: forall (r :: [Effect]). ToANF r => CoreExpr -> Eff r (Expr Var)

toANFCont :: forall (r :: [Effect]). ToANF r => CoreExpr -> ContT (Expr Var) (Eff r) (AExpr Var)

toANF' :: forall (r :: [Effect]). ToANF r => CoreExpr -> (AExpr Var -> Eff r (Expr Var)) -> Eff r (Expr Var)

Convert a Core expression to an atomic expression in ANF, accepting a continuation to handle the rest of the expression

toANFRec :: forall (r :: [Effect]). ToANF r => Expr Var -> (CExpr Var -> ContT (Expr Var) (Eff r) (Expr Var)) -> Eff r (Expr Var)

fromANF :: Expr Var -> CoreExpr

fromANFCExpr :: CExpr Var -> CoreExpr

fromANFAtom :: AExpr Var -> Expr Var