Elara.Core.TypeCheck

Description

Type *checking* for the Core language. This functions as a sanity check to make sure that any optimisations haven't broken the structure of the program to form an invalid program

It doesn't do any inference! As Core is already typed, it just checks that the types are consistent

Documentation

data TypeCheckError

Constructors

UnknownVariable Var (Set (UnlocatedVarRef Text))

CoreTypeMismatch

Fields

expected :: Type

actual :: Type

source :: (CoreExpr, CoreExpr)

errorCallStack :: CallStack

CoreTypeMismatchIncompleteExpected

Fields

incompleteExpected :: Text

actual :: Type

source :: (CoreExpr, CoreExpr)

UnificationError CoreExpr CoreExpr

InfiniteType Var CoreExpr

OccursCheck Var CoreExpr

PatternMatchMissingBinders

Fields

alt :: AltCon

altType :: Type

providedBinders :: [Var]

expr :: CoreExpr

newtype TcState

Constructors

TcState

Fields

scope :: Set (UnlocatedVarRef Text)

The Var already holds the variable's type so we don't need to track that. However we do need to track scoping, as an optimisation could pull a variable out of scope

addToScope :: Var -> TcState -> TcState

isInScope :: Var -> TcState -> Bool

typeCheckCoreModule :: forall (r :: [Effect]). (Error TypeCheckError :> r, StructuredDebug :> r) => CoreModule (Bind Var Expr) -> Eff r ()

varType :: Var -> Type

typeCheck :: forall (r :: [Effect]). (Error TypeCheckError :> r, State TcState :> r, StructuredDebug :> r, HasCallStack) => Expr Var -> Eff r Type

typeCheckC :: forall (r :: [Effect]). (HasCallStack, Error TypeCheckError :> r, State TcState :> r, StructuredDebug :> r, HasCallStack) => CExpr Var -> Eff r Type

typeCheckLit :: Literal -> Type

typeCheckA :: forall (r :: [Effect]). (Error TypeCheckError :> r, State TcState :> r, StructuredDebug :> r) => AExpr Var -> Eff r Type

equalUnderSubst :: Type -> Type -> Bool

Relation that defines 2 types as equal iff they are equal under a substitution of type variables For example forall a. a and forall b. b are equal in this relation, but forall a b. a -> b and forall a b. b -> a are not equal

equalUnderSubst' :: Type -> Type -> Bool

generalize :: Type -> Type