Elara.Data.Kind.Infer

Description

Performs kind inference on a type. Algorithm inspired by https://gilmi.me/blog/post/2023/09/30/kind-inference

> infer "type alias Identity a = a"

"k -> k"

> infer "type Maybe a = Either a Unit"

Type -> Type

> infer "type Either a b = Left a | Right b"

Type -> Type -> Type

> infer "type Proxy a = Proxy"

k -> Type

TODO: this is extremely basic and needs some nicer error messages, no support for polykinds or higher-kinded types yet

Documentation

data KindInferError

Constructors

UnknownKind (Qualified TypeName) (Map (Qualified TypeName) ElaraKind)

UnboundVar (Unique LowerAlphaName) (Map (Either (Qualified TypeName) (Unique LowerAlphaName)) KindVar)

CannotUnify ElaraKind ElaraKind ConstraintOrigin

NotFunctionKind ElaraKind

OccursCheckFailed KindVar ElaraKind

inferKind

Arguments

:: forall (r :: [Effect]). (HasCallStack, State InferState :> r, Error KindInferError :> r, Error SomeReportableError :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Rock Query :> r, StructuredDebug :> r, UniqueGen :> r, FileSystem :> r, Concurrent :> r)

=> Qualified TypeName
The name of the type being inferred

-> [Located (Unique LowerAlphaName)]
The type variables of the type being inferred

-> ShuntedTypeDeclaration
The type declaration to infer the kind of

-> Eff r (ElaraKind, KindedTypeDeclaration)

inferTypeKind :: forall (r :: [Effect]). KindInfer r => ShuntedType -> Eff r KindedType

initialInferState :: InferState

lookupKindVarMaybe :: forall (r :: [Effect]). KindInfer r => Either (Qualified TypeName) (Unique LowerAlphaName) -> Eff r (Maybe KindVar)

lookupNameKindVar :: forall (r :: [Effect]). (State InferState :> r, UniqueGen :> r, Error KindInferError :> r, QueryEffects r, Rock Query :> r) => Qualified TypeName -> ConstraintOrigin -> Eff r KindVar

preRegisterType :: forall (r :: [Effect]). (State InferState :> r, UniqueGen :> r, HasCallStack) => Qualified TypeName -> Eff r ()

Pass 1: Register a type name with a fresh kind variable (the "hole").

data InferState