Elara.Data.TopologicalGraph

Description

A generic, graph data structure that supports topological sorting.

Documentation

class Ord (Key a) => HasDependencies a where

Minimal complete definition

(keys | key), dependencies

Associated Types

type Key a

Methods

keys :: a -> NonEmpty (Key a)

key :: a -> Key a

dependencies :: a -> [Key a]

data TopologicalGraph a

Constructors

TopologicalGraph

Fields

_moduleGraph :: Graph

_nodeFromVertex :: Vertex -> (a, Key a, [Key a])

_vertexFromKey :: Key a -> Maybe Vertex

moduleGraph :: Lens' (TopologicalGraph a) Graph

nodeFromVertex :: Lens' (TopologicalGraph a) (Vertex -> (a, Key a, [Key a]))

vertexFromKey :: Lens' (TopologicalGraph a) (Key a -> Maybe Vertex)

mapGraph :: HasDependencies m => (a -> m) -> TopologicalGraph a -> TopologicalGraph m

traverseGraph :: (HasDependencies m, Applicative f) => (a -> f m) -> TopologicalGraph a -> f (TopologicalGraph m)

Traverse a graph. Due to the 'HasDependencies m' constraint we can't derive Traversable

traverseGraph_ :: Applicative f => (a -> f b) -> TopologicalGraph a -> f ()

traverseGraphTopologically :: (HasDependencies m, Applicative f) => (a -> f m) -> TopologicalGraph a -> f (TopologicalGraph m)

Traverse a graph in topological order

traverseGraphTopologically_ :: Applicative f => (a -> f m) -> TopologicalGraph a -> f ()

Traverse a graph in topological order, ignoring the results

traverseGraphRevTopologically :: (HasDependencies m, Applicative f) => (a -> f m) -> TopologicalGraph a -> f (TopologicalGraph m)

Traverse a graph in rev topological order

traverseGraphRevTopologically_ :: Applicative f => (a -> f m) -> TopologicalGraph a -> f ()

Traverse a graph in reverse topological order, ignoring the results

genericGraphTraverse :: (HasDependencies a, Applicative f) => (Graph -> [Vertex]) -> (m -> f a) -> TopologicalGraph m -> f (TopologicalGraph a)

genericGraphTraverse_ :: Applicative f => (Graph -> [Vertex]) -> (m -> f a) -> TopologicalGraph m -> f ()

createGraph :: HasDependencies a => [a] -> TopologicalGraph a

createEdge :: HasDependencies a => a -> NonEmpty (a, Key a, [Key a])

removeNode :: HasDependencies a => Key a -> TopologicalGraph a -> TopologicalGraph a

addNode :: HasDependencies a => a -> TopologicalGraph a -> TopologicalGraph a

replaceNode :: HasDependencies a => a -> TopologicalGraph a -> TopologicalGraph a

replaceNodeAt :: HasDependencies a => Key a -> a -> TopologicalGraph a -> TopologicalGraph a

allEntries :: TopologicalGraph m -> [m]

allEntriesTopologically :: Ord m => TopologicalGraph m -> [m]

allEntriesRevTopologically :: Ord m => TopologicalGraph m -> [m]

Get all entries in reverse topological order, but without duplicates

dependenciesOf :: Key a -> TopologicalGraph a -> [Key a]

moduleFromName :: Key a -> TopologicalGraph a -> Maybe a