Elara.Desugar

Documentation

type Desugar a = Eff DesugarPipelineEffects a

type DesugarPipelineEffects = '[State DesugarState, Error DesugarError, StructuredDebug]

newtype DesugarState

Constructors

DesugarState

Fields

_partialDeclarations :: Map (IgnoreLocation Name) PartialDeclaration

partialDeclarations :: Iso' DesugarState (Map (IgnoreLocation Name) PartialDeclaration)

getDesugaredModule :: ModuleName -> Eff (ConsQueryEffects '[Error DesugarError, Rock Query]) (Module SourceRegion Desugared)

desugar :: Module SourceRegion Frontend -> Desugar (Module SourceRegion Desugared)

desugarDeclarations :: Located ModuleName -> [Declaration SourceRegion Frontend] -> Desugar [Declaration SourceRegion Desugared]

assertPartialNamesEqual :: Eq a => (PartialDeclaration, Located a) -> (PartialDeclaration, Located a) -> Desugar ()

mergeAnnotations :: Maybe [Annotation SourceRegion Desugared] -> Maybe [Annotation SourceRegion Desugared] -> Desugar [Annotation SourceRegion Desugared]

mergePartials :: PartialDeclaration -> PartialDeclaration -> Desugar PartialDeclaration

resolvePartialDeclaration :: PartialDeclaration -> Desugar (DeclarationBody SourceRegion Desugared)

genPartials :: [Declaration SourceRegion Frontend] -> Desugar ()

completePartials :: Located ModuleName -> Desugar [Declaration SourceRegion Desugared]

desugarAnnotation :: Annotation SourceRegion Frontend -> Desugar (Annotation SourceRegion Desugared)

desugarAnnotationArg :: AnnotationArg SourceRegion Frontend -> Desugar (AnnotationArg SourceRegion Desugared)

desugarTypeDeclaration :: TypeDeclaration SourceRegion Frontend -> Desugar (TypeDeclaration SourceRegion Desugared)

desugarType :: Type SourceRegion Frontend -> Desugar (Type SourceRegion Desugared)

desugarExpr :: Expr SourceRegion Frontend -> Desugar (Expr SourceRegion Desugared)

desugarExprExtension :: FrontendExpressionExtension SourceRegion -> Desugar (Expr' SourceRegion Desugared)

desugarBinaryOperatorExtension :: BinaryOperatorExtension SourceRegion Frontend -> Desugar (BinaryOperatorExtension SourceRegion Desugared)

desugarBinaryOperator :: BinaryOperator SourceRegion Frontend -> Desugar (BinaryOperator SourceRegion Desugared)

desugarPattern :: Pattern SourceRegion Frontend -> Desugar (Pattern SourceRegion Desugared)

foldLambda :: [Pattern SourceRegion Desugared] -> Expr SourceRegion Desugared -> Expr SourceRegion Desugared

Lambdas need quite a lot of desugaring - they need to be unfolded into a series of nested lambdas, and then each pattern needs to be converted into a match expression.

For example, a (b, c) 1 -> e becomes a -> b_1 -> c_1 -> match b_1 with (b, c) -> match 1 with 1 -> e

However, converting the matches would require renaming, and we're not able to do that yet. Instead, we unfold the lambda, but keep the patterns, and the renamer handles the match conversion.

Orphan instances

PhaseCoerce (Exposing SourceRegion Frontend) (Exposing SourceRegion Desugared)

Instance details

Methods

phaseCoerce :: Exposing SourceRegion Frontend -> Exposing SourceRegion Desugared

PhaseCoerce (Exposition SourceRegion Frontend) (Exposition SourceRegion Desugared)

Instance details

Methods

phaseCoerce :: Exposition SourceRegion Frontend -> Exposition SourceRegion Desugared

PhaseCoerce (Import SourceRegion Frontend) (Import SourceRegion Desugared)

Instance details

Methods

phaseCoerce :: Import SourceRegion Frontend -> Import SourceRegion Desugared

PhaseCoerce (Import' SourceRegion Frontend) (Import' SourceRegion Desugared)

Instance details

Methods

phaseCoerce :: Import' SourceRegion Frontend -> Import' SourceRegion Desugared