Elara.Desugar.Common

Documentation

data PartialDeclaration

A partial declaration stores a desugared part of a declaration This allows merging of declarations with the same name For example, the code def a : Int ... ... let a = 5 is legal, and the 2 parts of the declaration need to be merged

Firstly, we create a JustDef after seeing the def line, then we merge this with a JustLet after seeing the let line to create a Both declaration, which is then resolved to a Desugared.Declaration'

Constructors

JustDef
A partial declaration with just a def line

Fields

(TaggedLocate 'VarNode SourceRegion VarName)

Name of the declaration

(NodeLoc 'DeclNode SourceRegion)

The *overall* region of the declaration, not just the body!

DesugaredType

(Maybe [Annotation SourceRegion Desugared])

JustLet

Fields

(TaggedLocate 'VarNode SourceRegion VarName)

Name of the declaration

(NodeLoc 'DeclNode SourceRegion)

The *overall* region of the *entire* declaration, not just the body!

DesugaredExpr

The desugared expression of the let binding

(Maybe [Annotation SourceRegion Desugared])

Any annotations on the let binding

AllDecl (TaggedLocate 'VarNode SourceRegion VarName) (NodeLoc 'DeclNode SourceRegion) DesugaredType DesugaredExpr [Annotation SourceRegion Desugared]
A def *and* let declaration that have been merged together, but not yet resolved to a Declaration

Immediate (Declaration SourceRegion Desugared)

partialDeclarationSourceRegion :: PartialDeclaration -> NodeLoc 'DeclNode SourceRegion

getPartialName :: PartialDeclaration -> TaggedLocate 'VarNode SourceRegion Name