Elara.Desugar.Error

Documentation

data DesugarError

Constructors

DefWithoutLet DesugaredType

InfixWithoutDeclaration (Located Name) SourceRegion (ValueDeclAnnotations 'Desugared)

DuplicateDeclaration PartialDeclaration PartialDeclaration

PartialNamesNotEqual PartialDeclaration PartialDeclaration

TuplePatternTooShort FrontendPattern

data PartialDeclaration

A partial declaration stores a desugared part of a declaration This allows merging of declarations with the same name For example, the code def a : Int ... ... let a = 5 is legal, and the 2 parts of the declaration need to be merged

Firstly, we create a JustDef after seeing the def line, then we merge this with a JustLet after seeing the let line to create a Both declaration, which is then resolved to a Desugared.Declaration'

Constructors

JustDef
A partial declaration with just a def line

Fields

(Located VarName)

Name of the declaration

SourceRegion

The *overall* region of the declaration, not just the body!

DesugaredType

(Maybe (ValueDeclAnnotations 'Desugared))

JustLet (Located VarName) SourceRegion DesugaredExpr (Maybe (ValueDeclAnnotations 'Desugared))

AllDecl (Located VarName) SourceRegion DesugaredType DesugaredExpr (ValueDeclAnnotations 'Desugared)

Immediate Name DesugaredDeclarationBody

partialDeclarationSourceRegion :: PartialDeclaration -> SourceRegion