Elara.Error.Effect

Documentation

data DiagnosticWriter t (a :: Type -> Type) b where

Essentially a very specialised Writer effect for Diagnostics, but safer | Because the Semigroup instance for Diagnostics is a little funky, there's a chance of accidentally overwriting the | entire FileMap, so this effect is designed to prevent that, and also prevent the state / writer boilerplate as an added bonus :)

Constructors

WriteDiagnostic :: forall t (a :: Type -> Type). Diagnostic t -> DiagnosticWriter t a ()

WriteReport :: forall t (a :: Type -> Type). Report t -> DiagnosticWriter t a ()

AddFile :: forall t (a :: Type -> Type). FilePath -> String -> DiagnosticWriter t a ()

addFile :: forall t (es :: [Effect]). (HasCallStack, DiagnosticWriter t :> es) => FilePath -> String -> Eff es ()

Perform the operation AddFile.

writeReport :: forall t (es :: [Effect]). (HasCallStack, DiagnosticWriter t :> es) => Report t -> Eff es ()

Perform the operation WriteReport.

writeDiagnostic :: forall t (es :: [Effect]). (HasCallStack, DiagnosticWriter t :> es) => Diagnostic t -> Eff es ()

Perform the operation WriteDiagnostic.

execDiagnosticWriter :: forall t (r :: [(Type -> Type) -> Type -> Type]) a. Eff (DiagnosticWriter t ': r) a -> Eff r (Diagnostic t)

evalDiagnosticWriter :: forall t (r :: [(Type -> Type) -> Type -> Type]) a. Eff (DiagnosticWriter t ': r) a -> Eff r a

runDiagnosticWriter :: forall t (r :: [(Type -> Type) -> Type -> Type]) a. Eff (DiagnosticWriter t ': r) a -> Eff r (Diagnostic t, a)