Elara.Error.Internal

Description

Module for internal compiler errors. Unlike the stage-based errors, these errors indicate that something has gone wrong internally, eg an invariant has been violated.

Documentation

data InternalError

Constructors

RequiredDeclNotFound (Qualified Name)

DuplicateDeclAfterDesugar ModuleName Name

InvalidAnnotationArrangement
When we parse an annotation at compile time but its arrangement is invalid