Elara.JVM.Error

Documentation

data JVMLoweringError

Errors that can occur during JVM lowering and code generation

Constructors

LambdaBinderNotLocal Var
Expected a local variable in a lambda, but got something else

MoreLambdasThanTypeArgs Text Int Int
A data constructor has more lambda abstractions than type parameters

DataConArityMismatch Text Int Int
A data constructor was applied to the wrong number of arguments

NotAFunctionalType Type
Expected a functional type, but got something else

UnsupportedType Type
A type constructor application was encountered that we don't know how to lower

LocalVariableNotFound Var
A variable was encountered that isn't bound in the current scope

InvokeStaticLocal Text
Tried to invoke a static method as a local variable

ClassNotFound Text
Unresolved reference to a class

MethodNotFound Text Text MethodDescriptor
Unresolved reference to a method

FieldNotFound Text Text Type
Unresolved reference to a field

GlobalVarInPattern (Qualified Text)
A global variable was found in a pattern match (not supported yet)

UnsupportedLiteral Literal
A literal was encountered that we don't know how to lower

TypeVarInPattern
A type variable was found in a pattern match (not supported)

OverApplicationOf (Text, MethodDescriptor) Int Int
A callable was applied to too many arguments

UnsupportedExpressionType CoreExpr
An expression of an unsupported type was encountered

AppOfNonFunction Text
A non-function was applied to an argument

MethodTooManyLocals Int

MethodTooManyStack Int

H2JVMError StackMapError