Elara.JVM.Lower.Match

Documentation

bindPatternVars

Arguments

:: forall (r :: [Effect]). InnerLower r

=> (Expr, ClassInfoType)
The scrutinee expression and its type

-> Alt Var
The pattern alternative

-> Eff r ()

Emits instructions to bind pattern variables from a pattern match alternative. For data constructor patterns, it generates field accesses to extract the fields from the scrutinee expression and assigns them to the corresponding pattern variables.

type ExprLoweringFn = forall (r :: [Effect]). InnerLower r => CoreExpr -> Eff r Expr

Simple way to break the circular dependency between lowering expressions and lowering matches

lowerMatch :: forall (r :: [Effect]). InnerLower r => ExprLoweringFn -> Type -> CoreExpr -> Maybe Var -> [Alt Var] -> Eff r Expr

matchCond :: Expr -> Alt Var -> Expr