Elara.Lexer.Token

Documentation

type Lexeme = Located Token

type TokPosition = RealPosition

data Token

Constructors

TokenSemicolon
;

TokenComma
,

TokenDot
.

TokenColon
:

TokenDoubleColon
::

TokenEquals
=

TokenBackslash
\

TokenPipe
|

TokenLeftParen
(

TokenRightParen
)

TokenLeftBrace
{

TokenRightBrace
}

TokenLeftBracket
[

TokenRightBracket
]

TokenLeftArrow
<-

TokenRightArrow

>

TokenDoubleRightArrow
>

TokenAt
@

TokenBacktick
`

TokenInt Integer

TokenFloat Double

TokenChar Char

TokenString Text
Keywords

TokenLet

TokenDef

TokenIn

TokenIf

TokenThen

TokenElse

TokenCase

TokenOf

TokenData

TokenType

TokenAlias

TokenModule

TokenImport

TokenAs

TokenQualified

TokenExposing

TokenWhere

TokenForall

TokenClass

TokenInstance

TokenDeriving

TokenMatch

TokenWith

TokenInfixL

TokenInfixR

TokenInfix

TokenVariableIdentifier Text
Variable Identifiers

TokenConstructorIdentifier Text

TokenOperatorIdentifier Text
Other

TokenQVariableIdentifier (ModuleName, Text)

TokenQConstructorIdentifier (ModuleName, Text)

TokenQOperatorIdentifier (ModuleName, Text)

TokenUnderscore

TokenIndent

TokenDedent

TokenLineSeparator

TokenEOF

tokenRepr :: Token -> Text

unsafeTokenText :: Token -> Text

unsafeTokenInt :: Token -> Integer

unsafeTokenFloat :: Token -> Double

isIndent :: Token -> Bool

tokenEndsExpr :: Token -> Bool

Returns True if the token ends an expression. This is used to determine whether line breaks should be treated as semicolons If the token doesn't end an expression, then a line break is skipped. If the token does end an expression, then a line break is treated as a semicolon.