Elara.Lexer.Utils

Documentation

type LexMonad a = Eff '[State ParseState, Error LexerError, StructuredDebug] a

data ParseState

data AlexInput

data LexerError

Constructors

TooMuchIndentation
When an element is indented more than expected

Fields

IndentInfo

The expected indentation

(Maybe IndentInfo)

The potential further indentation

Int

The actual indentation

ParseState

The current state of the lexer

UnterminatedStringLiteral ParseState

GenericAlexError AlexInput

getPosition :: Int -> LexMonad TokPosition

stringBuf :: Lens' ParseState Text

lexSC :: Lens' ParseState Int

commentDepth :: Lens' ParseState Int

pendingPosition :: Lens' ParseState TokPosition

input :: Lens' ParseState AlexInput

pendingTokens :: Lens' ParseState [Lexeme]

position :: Lens' AlexInput RealPosition

rest :: Lens' AlexInput Text

cleanIndentation :: LexMonad [Lexeme]

createRegionStartingAt :: TokPosition -> LexMonad RealSourceRegion

emitAt :: Token -> SourceRegion -> LexMonad (Maybe Lexeme)

createRegion :: TokPosition -> TokPosition -> LexMonad RealSourceRegion

splitQualName :: Text -> (ModuleName, Text)

Splits a qualified name into the qualifier and the name. Throws an error if the name is not qualified.

Examples:

>>> splitQualName "Hello.world"
(ModuleName ("Hello" :| []),"world")

>>> splitQualName "A.B.C"
(ModuleName ("A" :| ["B"]),"C")

>>> splitQualName "Prelude..+"
(ModuleName ("Prelude" :| []),".+")

>>> splitQualName "A.!."
(ModuleName ("A" :| []),"!.")

startWhite :: Int -> Text -> LexMonad (Maybe Lexeme)

alexGetByte :: AlexInput -> Maybe (Word8, AlexInput)

Read a byte from the input

initialLexState :: FilePath -> Text -> ParseState

triggerIndentLayout :: LexMonad ()

triggerBlockLayout :: LexMonad ()

checkBlockLayout :: Int -> LexMonad ()

If a (block) layout is expected, check the current token's indentation against the top of the stack, and insert an INDENT token if necessary