Elara.Parse.Combinators

Documentation

sepBy1' :: Parser a -> Parser sep -> Parser (NonEmpty a)

Safe version of sepBy1 that backtracks if the parser after the separator fails. Could also be considered a lazy version of [sepBy1]

sepEndBy1' :: Parser a -> Parser sep -> Parser (NonEmpty a)

Safe version of sepEndBy1 that backtracks if the parser after the separator fails. Could also be considered a lazy version of sepBy1'

liftedBinary :: (JoinKinds k1 A_Lens k2, Monad m, Is k2 A_Getter, Is (ReversedOptic k1) A_Getter, ReversibleOptic k1) => m t -> (t -> a1 -> a1 -> a2) -> Optic k1 NoIx a1 a1 (Located a2) (Located a2) -> m (a1 -> a1 -> a1)