Elara.Parse.Error

Documentation

data ElaraParseError

Constructors

KeywordUsedAsName (Located (MaybeQualified VarName))

EmptyRecord SourceRegion

EmptyLambda SourceRegion

InfixPrecTooHigh (Located Integer)

parseErrorSources :: ElaraParseError -> [SourceRegion]

newtype WParseErrorBundle e m

Constructors

WParseErrorBundle

Fields

unWParseErrorBundle :: ParseErrorBundle e m

diagnosticFromBundle

Arguments

:: forall msg s e. (Token s ~ Lexeme, e ~ ElaraParseError, IsString msg, HasHints e msg, ShowErrorComponent e, VisualStream s, TraversableStream s)

=> (ParseError s e -> Bool)
How to decide whether this is an error or a warning diagnostic

-> Maybe msg
An optional error code

-> msg
The error message of the diagnostic

-> Maybe [Note msg]
Default hints when trivial errors are reported

-> ParseErrorBundle s ElaraParseError
The bundle to create a diagnostic from

-> Diagnostic msg

This is a slightly modified version of errorDiagnosticFromBundle from the diagnose package. It adds the ability to highlight a region of the source code rather than a single point for error highlighting.

both :: (a -> b) -> (a, a) -> (b, b)