Elara.Parse.Expression

Documentation

data ExpressionGrammar

Data type putting all expression parsers together to avoid deadlocks in CAF initialisation.

Constructors

ExpressionGrammar

Fields

pElement :: Parser FrontendExpr

pExpression :: Parser FrontendExpr

the main expression parser

locatedExpr :: Parser FrontendExpr' -> Parser FrontendExpr

letPreamble :: ExpressionGrammar -> Parser (TaggedLocate 'VarNode SourceRegion VarName, [FrontendPattern], FrontendExpr)

expression :: ExpressionGrammar -> Parser FrontendExpr

The main recursive expression parser. Adds function application, cons, and binary operators on top of the basic terms.

element :: ExpressionGrammar -> Parser FrontendExpr

A top level element, comprised of either an expression or a let declaration Note that top level let declarations are not parsed here, but in the Elara.Parse.Declaration module.

reservedWords :: Set Text

Reserved words, used to backtrack accordingly