Elara.Parse.Grammar

Description

Ties the knot for the expression grammar, allowing for mutually recursive parsers without risk of deadlocks 🥲

Documentation

exprParser :: Parser FrontendExpr

element :: Parser FrontendExpr

letPreambleParser :: Parser (TaggedLocate 'VarNode SourceRegion VarName, [FrontendPattern], FrontendExpr)