Elara.Parse.Primitives

Documentation

type Parser = Parsec ElaraParseError TokenStream

fmapLocated :: (Located a -> b) -> Parser a -> Parser b

located :: Parser a -> Parser (Located a)

A parser that records the location information of the tokens it consumes. TODO this is not going to perform very well as it's O(n) in the total number of input tokens A future solution will be to store the number of tokens consumed in the TokenStream and use that to calculate the spanning region, but that's effort at the moment.

optionallyInParens :: Parser a -> Parser a

inParens :: Parser a -> Parser a

inBraces :: Parser a -> Parser a

commaSeparated :: Parser a -> Parser [a]

oneOrCommaSeparatedInParens :: Parser a -> Parser [a]

token :: Token -> Parser Token

token_ :: Token -> Parser ()

withPredicate :: (t -> Bool) -> (t -> ElaraParseError) -> Parser t -> Parser t

satisfyMap :: (Token -> Maybe a) -> Parser a

lexeme :: Token -> Parser Lexeme

locatedTokens' :: NonEmpty Token -> Parser SourceRegion

ignoringIndents :: Parser a -> Parser a

Temporarily puts the parser in a state where it ignores indents this is useful for parsing things that are truly context-free, such as type declarations where the indentation level is not significant.