Elara.Pipeline

Description

Each stage of AST processing can be interpreted as part of an overall pipeline of effects in the Sem monad. This acts as the entrypoint to the stage, bringing each stage into a common abstraction

Documentation

runLogToStdoutAndFile :: forall (es :: [Effect]) a. IOE :> es => Eff (Log (Doc AnsiStyle) ': es) a -> Eff es a