Elara.Query

Description

Documentation

type WithRock (effects :: [Effect]) = Rock Query ': effects

data ASTQueryType

Constructors

QModule

QDecl

QReqDecl

QCtor

QAnn

QAnnType

class (Typeable ast, ElaraPhase ast) => RunPhase ast where

Associated Types

type ASTQueryEffects ast (q :: ASTQueryType) :: [Effect]

type ASTQueryEffects ast (q :: ASTQueryType) = StandardQueryError ast

Methods

getModuleByName :: ModuleName -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects ast 'QModule))) (Module SourceRegion ast)

getDeclarationByName :: Qualified Name -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects ast 'QDecl))) (Maybe (Declaration SourceRegion ast))

getRequiredDeclarationByName :: Qualified Name -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects ast 'QReqDecl))) (Declaration SourceRegion ast)

getConstructorDeclaration :: ConstructorOccurrence ast SourceRegion -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects ast 'QCtor))) (Declaration SourceRegion ast)

getDeclarationAnnotations :: Qualified Name -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects ast 'QAnn))) [Annotation SourceRegion ast]

getDeclarationAnnotationsOfType :: (Qualified Name, Qualified TypeName) -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects ast 'QAnnType))) [Annotation SourceRegion ast]

data Query (es :: [Effect]) a where

Constructors

GetCompilerSettings :: Query (Rock Query ': MinimumQueryEffects) CompilerSettings
Query to get the compiler settings

InputFiles :: Query (Rock Query ': ConsMinimumQueryEffects '[FileSystem]) (HashSet FilePath)
Query to get all the required input files to be passed to the compiler

ModuleIndex :: Query (Rock Query ': ConsMinimumQueryEffects '[FileSystem]) ModuleIndex
Query to get the module index (bidirectional mapping between file paths and module names)

GetFileContents :: FilePath -> Query (Rock Query ': ConsMinimumQueryEffects '[FileSystem, Error ReadFileError]) FileContents
Query to get the contents of a specific file

ModulePath :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error ModulePathError]) [Char]
Query to get the file path of a module

LexedFile :: FilePath -> Query (Rock Query ': ConsQueryEffects '[Error LexerError]) [Lexeme]
Query to get the lexed tokens of a specific file

ParsedFile :: FilePath -> Query (Rock Query ': ConsQueryEffects '[Error (WParseErrorBundle TokenStream ElaraParseError)]) (Module SourceRegion Frontend)
Query to get the parsed module from a file's contents and lexed tokens

ParsedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error (WParseErrorBundle TokenStream ElaraParseError)]) (Module SourceRegion Frontend)
Query to get a parsed module by module name

DesugaredModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error DesugarError]) (Module SourceRegion Desugared)

ModuleByName :: forall ast. (RunPhase ast, Typeable ast) => ModuleName -> Query (Rock Query ': ConsQueryEffects (ASTQueryEffects ast 'QModule)) (Module SourceRegion ast)

DeclarationByName :: forall ast. (RunPhase ast, Typeable ast) => Qualified Name -> Query (Rock Query ': ConsQueryEffects (ASTQueryEffects ast 'QDecl)) (Maybe (Declaration SourceRegion ast))

RequiredDeclarationByName :: forall ast. (RunPhase ast, Typeable ast) => Qualified Name -> Query (Rock Query ': ConsQueryEffects (ASTQueryEffects ast 'QReqDecl)) (Declaration SourceRegion ast)

ConstructorDeclaration :: forall ast. (RunPhase ast, Typeable ast, Show (ConstructorOccurrence ast SourceRegion), Ord (ConstructorOccurrence ast SourceRegion), Hashable (ConstructorOccurrence ast SourceRegion)) => ConstructorOccurrence ast SourceRegion -> Query (Rock Query ': ConsQueryEffects (ASTQueryEffects ast 'QCtor)) (Declaration SourceRegion ast)

DeclarationAnnotations :: forall ast. (RunPhase ast, Typeable ast) => Qualified Name -> Query (Rock Query ': ConsQueryEffects (ASTQueryEffects ast 'QAnn)) [Annotation SourceRegion ast]

DeclarationAnnotationsOfType :: forall ast. (RunPhase ast, Typeable ast) => (Qualified Name, Qualified TypeName) -> Query (Rock Query ': ConsQueryEffects (ASTQueryEffects ast 'QAnnType)) [Annotation SourceRegion ast]

GetOpInfo :: IgnoreLocVarRef Name -> Query (Rock Query ': ConsQueryEffects '[Writer (Set ShuntWarning), Error ShuntError]) (Maybe OpInfo)

GetOpTableIn :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Map (IgnoreLocVarRef Name) OpInfo)

FreeVarsOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (HashSet (Qualified VarName))

ReachableSubgraphOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) ReachableSubgraph

GetSCCsOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) [SCC (Qualified VarName)]

SCCKeyOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) SCCKey

TypeCheckedExpr :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Expr SourceRegion Typed)

TypeOf :: forall loc. loc ~ SourceRegion => TypeEnvKey loc -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Type loc)

InferSCC :: SCCKey -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Map (Qualified VarName) (Polytype SourceRegion))

KindOf :: Qualified TypeName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe KindVar)

GetTypeAlias
Get Information about a type alias

Fields

:: Qualified TypeName

The name of the type alias

-> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe ([UniqueTyVar], Type SourceRegion))

The type alias's type variables and body, if it exists

GetCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

GetTyCon :: Qualified Text -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe TyCon)

GetDataCon :: Qualified TypeName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe DataCon)

TypeCheckedDeclaration :: Qualified Name -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Declaration SourceRegion Typed)

GetOptimisedCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

GetANFCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule (TopLevelBind Var))

GetClosureLiftedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error ClosureLiftError]) (CoreModule (TopLevelBind Var))

GetFinalisedCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

GetJVMIRModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error JVMLoweringError]) Module

GetJVMClassFiles :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error JVMLoweringError]) [ClassFile]

GetJVMClassBytes :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error JVMLoweringError, Error CodeConverterError]) [(FilePath, LByteString)]

tagQuery :: forall {w1 :: [Effect]} {w2}. Query w1 w2 -> Int

sameCtor :: HasCallStack => forall (es :: [Effect]) a b. Query es a -> Query es b -> GOrdering a b