Elara.Query.Effects

Description

Common module with query effects To avoid a circular import we don't include the Rock effect here even though it's commonly used sorry :)

Documentation

type MinimumQueryEffects = ConsMinimumQueryEffects ('[] :: [Effect])

type ConsMinimumQueryEffects (es :: [Effect]) = Memoise ': (Concurrent ': es)

The bare minimum effects that need to be present for any query, even a "pure" one

type HasMinimumQueryEffects (es :: [Effect]) = (Memoise :> es, Concurrent :> es)

type StandardQueryEffects = ConsQueryEffects ('[] :: [Effect])

type ConsQueryEffects (es :: [Effect]) = FileSystem ': (Error SomeReportableError ': (DiagnosticWriter (Doc AnsiStyle) ': (UniqueGen ': (StructuredDebug ': ConsMinimumQueryEffects es))))

Standard effects that almost every query will use

type QueryEffects (es :: [Effect]) = (FileSystem :> es, Memoise :> es, Concurrent :> es, Error SomeReportableError :> es, DiagnosticWriter (Doc AnsiStyle) :> es, UniqueGen :> es, StructuredDebug :> es)