Elara.Rename

Description

Renaming stage of compilation This stage handles: 1. Renaming all variables, types, and type variables, adding module qualification or unique suffixes to avoid name clashes 2. Desugaring any "first-class" pattern matches into normal match expressions (eg '[] -> 1' to 'x -> match x with [] -> 1') 3. Desugaring blocks into let-in chains (and monad operations soon), eg 'let y = 1; y + 1' to 'let y = 1 in y + 1' Note that until the monad operations are implemented, we can't fully remove blocks, as we have nothing to translate 'f x; g x' into

Documentation

renameExpr :: forall (r :: [Effect]). (InnerRename r, Reader (Maybe (Declaration SourceRegion Desugared)) :> r, Rock Query :> r) => DesugaredExpr -> Eff r RenamedExpr

type InnerRename (r :: [Effect]) = (State RenameState :> r, Error RenameError :> r, UniqueGen :> r, QueryEffects r, StructuredDebug :> r, Reader (Maybe (Module SourceRegion Desugared)) :> r, HasCallStack)

Orphan instances

RunPhase Renamed

Instance details

Associated Types

type ASTQueryEffects Renamed q

Instance details

Defined in Elara.Rename

type ASTQueryEffects Renamed q = StandardQueryError Renamed

Methods

getModuleByName :: ModuleName -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects Renamed 'QModule))) (Module SourceRegion Renamed)

getDeclarationByName :: Qualified Name -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects Renamed 'QDecl))) (Maybe (Declaration SourceRegion Renamed))

getRequiredDeclarationByName :: Qualified Name -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects Renamed 'QReqDecl))) (Declaration SourceRegion Renamed)

getConstructorDeclaration :: ConstructorOccurrence Renamed SourceRegion -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects Renamed 'QCtor))) (Declaration SourceRegion Renamed)

getDeclarationAnnotations :: Qualified Name -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects Renamed 'QAnn))) [Annotation SourceRegion Renamed]

getDeclarationAnnotationsOfType :: (Qualified Name, Qualified TypeName) -> Eff (WithRock (ConsQueryEffects (ASTQueryEffects Renamed 'QAnnType))) [Annotation SourceRegion Renamed]