Elara.Rules.Generic

Description

AST agnostic implementations of rules

Documentation

declarationName :: forall {k} (p :: k) loc. (ElaraPhase p, TopValueBinder p loc ~ Located (Qualified VarName), TopTypeBinder p loc ~ Located (Qualified TypeName)) => Declaration loc p -> Name

Helper to get the name from a declaration (works for phases with Located Qualified binders)

hasConstructor :: forall {k} (ast :: k). (ElaraPhase ast, ConstructorBinder ast SourceRegion ~ Located (Qualified TypeName)) => Located (Qualified TypeName) -> Declaration SourceRegion ast -> Bool

Check if a declaration contains a type declaration with the given constructor name

Orphan instances

(SupportsQuery 'QueryRequiredDeclarationByName ast, Typeable ast, DeclBodyExtension ast SourceRegion ~ Void) => SupportsQuery 'DeclarationAnnotations ast

Instance details

Associated Types

type QuerySpecificEffectsOf 'DeclarationAnnotations ast

Instance details

Defined in Elara.Rules.Generic

type QuerySpecificEffectsOf 'DeclarationAnnotations ast = StandardQueryError ast

Methods

query :: QueryArgsOf 'DeclarationAnnotations ast -> Eff (QueryEffectsOf 'DeclarationAnnotations ast) (QueryReturnTypeOf 'DeclarationAnnotations ast)

(Typeable ast, SupportsQuery 'DeclarationAnnotations ast, SupportsQuery 'QueryConstructorDeclaration ast, ElaraPhase ast, ConstructorOccurrence ast SourceRegion ~ Located (Qualified TypeName), TypeOccurrence ast SourceRegion ~ Located (Qualified TypeName), TopTypeBinder ast SourceRegion ~ Located (Qualified TypeName)) => SupportsQuery 'DeclarationAnnotationsOfType ast

Instance details

Associated Types

type QuerySpecificEffectsOf 'DeclarationAnnotationsOfType ast

Instance details

Defined in Elara.Rules.Generic

type QuerySpecificEffectsOf 'DeclarationAnnotationsOfType ast = StandardQueryError ast

Methods

query :: QueryArgsOf 'DeclarationAnnotationsOfType ast -> Eff (QueryEffectsOf 'DeclarationAnnotationsOfType ast) (QueryReturnTypeOf 'DeclarationAnnotationsOfType ast)

(HasMinimumQueryEffects (QueryEffectsOf 'QueryConstructorDeclaration ast), ConstructorOccurrence ast SourceRegion ~ Located (Qualified TypeName), ConstructorBinder ast SourceRegion ~ Located (Qualified TypeName), ElaraPhase ast, Subset (QueryEffectsOf 'QueryModuleByName ast) (QueryEffectsOf 'QueryConstructorDeclaration ast), SupportsQuery 'QueryModuleByName ast, Typeable ast, Rock Query :> QueryEffectsOf 'QueryConstructorDeclaration ast, QueryReturnTypeOf 'QueryConstructorDeclaration ast ~ Declaration SourceRegion ast, QueryReturnTypeOf 'QueryModuleByName ast ~ Module SourceRegion ast) => SupportsQuery 'QueryConstructorDeclaration ast

Instance details

Associated Types

type QuerySpecificEffectsOf 'QueryConstructorDeclaration ast

Instance details

Defined in Elara.Rules.Generic

type QuerySpecificEffectsOf 'QueryConstructorDeclaration ast = StandardQueryError ast

Methods

query :: QueryArgsOf 'QueryConstructorDeclaration ast -> Eff (QueryEffectsOf 'QueryConstructorDeclaration ast) (QueryReturnTypeOf 'QueryConstructorDeclaration ast)

(SupportsQuery 'QueryModuleByName ast, HasMinimumQueryEffects (QueryEffectsOf 'QueryDeclarationByName ast), Rock Query :> QueryEffectsOf 'QueryDeclarationByName ast, Typeable ast, ElaraPhase ast, TopValueBinder ast SourceRegion ~ Located (Qualified VarName), TopTypeBinder ast SourceRegion ~ Located (Qualified TypeName), Subset (QueryEffectsOf 'QueryModuleByName ast) (QueryEffectsOf 'QueryDeclarationByName ast), QueryReturnTypeOf 'QueryModuleByName ast ~ Module SourceRegion ast) => SupportsQuery 'QueryDeclarationByName ast

Instance details

Associated Types

type QuerySpecificEffectsOf 'QueryDeclarationByName ast

Instance details

Defined in Elara.Rules.Generic

type QuerySpecificEffectsOf 'QueryDeclarationByName ast = StandardQueryError ast

Methods

query :: QueryArgsOf 'QueryDeclarationByName ast -> Eff (QueryEffectsOf 'QueryDeclarationByName ast) (QueryReturnTypeOf 'QueryDeclarationByName ast)

(SupportsQuery 'QueryDeclarationByName ast, Subset (QueryEffectsOf 'QueryDeclarationByName ast) (QueryEffectsOf 'QueryRequiredDeclarationByName ast), HasMinimumQueryEffects (QueryEffectsOf 'QueryRequiredDeclarationByName ast), Rock Query :> QueryEffectsOf 'QueryRequiredDeclarationByName ast, Typeable ast) => SupportsQuery 'QueryRequiredDeclarationByName ast

Instance details

Associated Types

type QuerySpecificEffectsOf 'QueryRequiredDeclarationByName ast

Instance details

Defined in Elara.Rules.Generic

type QuerySpecificEffectsOf 'QueryRequiredDeclarationByName ast = QuerySpecificEffectsOf 'QueryDeclarationByName ast

Methods

query :: QueryArgsOf 'QueryRequiredDeclarationByName ast -> Eff (QueryEffectsOf 'QueryRequiredDeclarationByName ast) (QueryReturnTypeOf 'QueryRequiredDeclarationByName ast)