Elara.Shunt

Description

This module performs "shunting", the process of rearranging binary operators in expressions to match their defined precedence and associativity. The main meat of this module is fixOperators, which does the actual rearranging of operators in expressions. The logic for this is based on https://stackoverflow.com/a/67992584/6272977, which was very helpful :).

Most of the other functions in this module are less interesting and mainly plumbing into the compiler, particularly a lot of boilerplate on traversing the AST.

Documentation

runGetOpInfoQuery :: SupportsQueries '['QueryDeclarationByName, 'DeclarationAnnotations, 'QueryConstructorDeclaration] 'Renamed => IgnoreLocVarRef Name -> Eff (ConsQueryEffects '[Writer (Set ShuntWarning), Error ShuntError, Rock Query]) (Maybe OpInfo)

Run the GetOpInfo query to get operator info for a given operator

runGetOpTableInQuery :: ModuleName -> Eff (ConsQueryEffects '[Rock Query]) OpTable

Run the GetOpTableIn query to get the operator table for a module

Orphan instances

IntrospectableAnnotations 'Shunted

Instance details

Methods

getAnnotations :: forall (sel :: ForSelector). Select ('Annotations sel) 'Shunted -> [Annotation 'Shunted]

SupportsQuery 'QueryModuleByName 'Shunted

Instance details

Associated Types

type QuerySpecificEffectsOf 'QueryModuleByName 'Shunted

Instance details

Defined in Elara.Shunt

type QuerySpecificEffectsOf 'QueryModuleByName 'Shunted = StandardQueryError 'Shunted

Methods

query :: QueryArgsOf 'QueryModuleByName 'Shunted -> Eff (QueryEffectsOf 'QueryModuleByName 'Shunted) (QueryReturnTypeOf 'QueryModuleByName 'Shunted)