Elara.ToCore

Description

Converts typed AST to Core

Documentation

data ToCoreError

Constructors

LetInTopLevel !TypedExpr

UnknownConstructor !(Located (Qualified TypeName)) CtorSymbolTable

UnknownPrimConstructor !(Qualified TypeName)

UnknownTypeConstructor !(Qualified Text) CtorSymbolTable

UnknownLambdaType !(Type SourceRegion)

UnsolvedTypeSnuckIn !(Type SourceRegion)

UnknownVariable !(Located (Qualified Name))

data CtorSymbolTable

Constructors

CtorSymbolTable

Fields

dataCons :: Map (Qualified Text) DataCon

tyCons :: Map (Qualified Text) TyCon

primCtorSymbolTable :: CtorSymbolTable

lookupCtor :: forall (r :: [Effect]). ToCoreC r => Located (Qualified TypeName) -> Eff r DataCon

registerCtor :: forall (r :: [Effect]). ToCoreC r => DataCon -> Eff r ()

registerTyCon :: forall (r :: [Effect]). ToCoreC r => TyCon -> Eff r ()

lookupPrimCtor :: forall (r :: [Effect]). ToCoreC r => Qualified TypeName -> Eff r DataCon

lookupTyCon :: forall (r :: [Effect]). (HasCallStack, ToCoreC r) => Qualified Text -> Eff r TyCon

runGetDataConQuery :: Qualified TypeName -> Eff (ConsQueryEffects '[Rock Query]) (Maybe DataCon)

runGetTyConQuery :: Qualified Text -> Eff (ConsQueryEffects '[Rock Query]) (Maybe TyCon)

type VariableTable = Map (Qualified Name) (Type SourceRegion)

type ToCoreEffects = '[State CtorSymbolTable, Error ToCoreError, UniqueGen, StructuredDebug]

type InnerToCoreEffects = '[State CtorSymbolTable, Error ToCoreError, UniqueGen, StructuredDebug]

type ToCoreC (r :: [Effect]) = (State CtorSymbolTable :> r, Error ToCoreError :> r, UniqueGen :> r, StructuredDebug :> r, QueryEffects r, Rock Query :> r)

type InnerToCoreC (r :: [Effect]) = (State CtorSymbolTable :> r, Error ToCoreError :> r, UniqueGen :> r, StructuredDebug :> r, QueryEffects r, Rock Query :> r)

runGetCoreModuleQuery :: ModuleName -> Eff (ConsQueryEffects '[Rock Query]) (CoreModule CoreBind)

sccToCore :: forall (r :: [Effect]). ToCoreC r => SCC (Qualified VarName, TypedExpr) -> Eff r (CoreDeclaration CoreBind)

buildModuleSCCs :: forall (r :: [Effect]). ToCoreC r => Module 'Typed -> Eff r [SCC (Qualified VarName)]

moduleToCore :: forall (r :: [Effect]). (HasCallStack, ToCoreC r) => Module 'Typed -> Eff r (CoreModule CoreBind)

mkTypeVar :: Select "TypeVar" 'Typed -> TypeVariable

polytypeToCore :: forall (r :: [Effect]). (HasCallStack, InnerToCoreC r) => Polytype SourceRegion -> Eff r Type

eitherTypeToCore :: forall (r :: [Effect]). (HasCallStack, InnerToCoreC r) => Type SourceRegion -> Eff r Type

typeToCore :: forall (r :: [Effect]). (HasCallStack, InnerToCoreC r) => Monotype SourceRegion -> Eff r Type

conToVar :: DataCon -> Var

toCore :: forall (r :: [Effect]). (HasCallStack, InnerToCoreC r) => TypedExpr -> Eff r CoreExpr

isRecursive :: forall {v1} {v2} {v3} {k} {l} {u} {v4} {a} {is :: IxList}. (ErrorUnlessOnlyOne v1 (Rep v1), Defined (Rep v2) (TypeError (('Text "| " ':<>: ('Text "No instance for " ':<>: (('Text "\8216" ':<>: 'ShowType (Generic v2)) ':<>: 'Text "\8217"))) ':$$: (('Text "| " ':<>: ((('Text "arising from a generic prism focusing on the " ':<>: (('Text "\8216" ':<>: 'ShowType "Var") ':<>: 'Text "\8217")) ':<>: 'Text " constructor of type ") ':<>: (('Text "\8216" ':<>: 'ShowType (Located v3)) ':<>: 'Text "\8217"))) ':$$: (('Text "| " ':<>: ('Text "in " ':<>: (('Text "\8216" ':<>: 'ShowType v2) ':<>: 'Text "\8217"))) ':$$: 'Text ""))) :: Constraint) (), ErrorUnless "Var" v2 (HasCtorP "Var" (Rep v2)), Is k A_Fold, JoinKinds A_Fold l k, GPlate v1 v1, GWrapped (Rep v1) (Rep v1) u v4, UnifyHead v1 v1, Field1 u v4 (Located v2) (Located v2), Generic v1, Generic v2, GAsConstructor "Var" (Rep v2) (Rep v2) (Located v3) (Located v3), Eq a) => a -> v1 -> Optic l is v3 v3 a a -> Bool

Check if a variable is recursive in an expression Takes the name of the variable, the expression, and a lens to the variable name (Local or Global usually)

desugarMatch :: forall (r :: [Effect]). (HasCallStack, InnerToCoreC r) => TypedExpr -> [(TypedPattern, TypedExpr)] -> Eff r CoreExpr

mkBindName :: forall (r :: [Effect]). InnerToCoreC r => TypedExpr -> Eff r Var

desugarBlock :: forall (r :: [Effect]). InnerToCoreC r => NonEmpty TypedExpr -> Eff r CoreExpr