Elara.TypeInfer

Documentation

type InferPipelineEffects (r :: [Effect]) = (StructuredDebug :> r, State InferState :> r, UniqueGen :> r, Error (UnifyError SourceRegion) :> r, Error TypeConvertError :> r, Error KindInferError :> r, Infer SourceRegion r)

runGetTypeCheckedModuleQuery :: SupportsQuery 'QueryModuleByName Shunted => ModuleName -> Eff (ConsQueryEffects '[Rock Query]) (Module SourceRegion Typed)

runTypeOfQuery :: (loc ~ SourceRegion, SupportsQuery 'QueryRequiredDeclarationByName Shunted, SupportsQuery 'QueryModuleByName Shunted) => TypeEnvKey loc -> Eff (ConsQueryEffects '[Rock Query]) (Type loc)

Run the TypeOf query to get the type of a term or data constructor

runKindOfQuery :: forall (r :: [(Type -> Type) -> Type -> Type]). SupportsQuery 'QueryModuleByName Shunted => Qualified TypeName -> Eff (ConsQueryEffects (Rock Query ': r)) (Maybe KindVar)

runGetTypeAliasQuery :: forall (r :: [(Type -> Type) -> Type -> Type]). SupportsQuery 'QueryModuleByName Shunted => Qualified TypeName -> Eff (ConsQueryEffects (Rock Query ': r)) (Maybe ([UniqueTyVar], Type SourceRegion))

seedConstructorsFor :: forall (r :: [Effect]). (SupportsQuery 'QueryModuleByName Shunted, HasCallStack, Rock Query :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Error SomeReportableError :> r, Error KindInferError :> r, Error TypeConvertError :> r, State (TypeEnvironment SourceRegion) :> r, State InferState :> r, Concurrent :> r, UniqueGen :> r, FileSystem :> r, StructuredDebug :> r) => ModuleName -> Eff r ()

runInferEffects :: forall (r :: [Effect]) a loc. (Pretty loc, QueryEffects r, Rock Query :> r, Eq loc, loc ~ SourceRegion) => Eff (InferEffectsCons loc (Error (UnifyError loc) ': (Error KindInferError ': (Error TypeConvertError ': r)))) a -> Eff r (a, Constraint loc)

runInferSCCQuery :: forall (r :: [(Type -> Type) -> Type -> Type]). SupportsQuery 'QueryRequiredDeclarationByName Shunted => SCCKey -> Eff (ConsQueryEffects (Rock Query ': r)) (Map (Qualified VarName) (Polytype SourceRegion))

seedSCC :: forall (r :: [Effect]). (SupportsQuery 'QueryRequiredDeclarationByName Shunted, DiagnosticWriter (Doc AnsiStyle) :> r, Rock Query :> r, Error SomeReportableError :> r, Error KindInferError :> r, Error TypeConvertError :> r, State (TypeEnvironment SourceRegion) :> r, State InferState :> r, StructuredDebug :> r, FileSystem :> r, UniqueGen :> r, Concurrent :> r) => SCC (Qualified VarName) -> Eff r ()

inferSCC :: forall (r :: [Effect]). (SupportsQuery 'QueryRequiredDeclarationByName Shunted, DiagnosticWriter (Doc AnsiStyle) :> r, Rock Query :> r, Error (InferError SourceRegion) :> r, Error (UnifyError SourceRegion) :> r, Error SomeReportableError :> r, State (TypeEnvironment SourceRegion) :> r, State (LocalTypeEnvironment SourceRegion) :> r, Reader ContextStack :> r, StructuredDebug :> r, FileSystem :> r, UniqueGen :> r, Concurrent :> r) => SCC (Qualified VarName) -> Eff r (Map (Qualified VarName) (Polytype SourceRegion))

runTypeCheckedExprQuery :: forall (r :: [(Type -> Type) -> Type -> Type]). Qualified VarName -> Eff (ConsQueryEffects (Rock Query ': r)) TypedExpr

runTypeCheckedDeclarationQuery :: forall (r :: [(Type -> Type) -> Type -> Type]). Qualified Name -> Eff (ConsQueryEffects (Rock Query ': r)) TypedDeclaration

declarationName :: Declaration SourceRegion Typed -> Name

Get the name from a typed declaration

collectConstructors :: ShuntedExpr -> [Qualified TypeName]

Collect all constructor references from a shunted expression

collectPatternConstructors :: ShuntedPattern -> [Qualified TypeName]

Collect constructor references from a shunted pattern

inferModule :: forall (r :: [Effect]). (InferPipelineEffects r, Infer SourceRegion r) => Module SourceRegion Shunted -> Eff r (Module SourceRegion Typed)

seedDeclaration :: forall (r :: [Effect]). (DiagnosticWriter (Doc AnsiStyle) :> r, Rock Query :> r, Error KindInferError :> r, Error SomeReportableError :> r, Error TypeConvertError :> r, State (TypeEnvironment SourceRegion) :> r, State InferState :> r, StructuredDebug :> r, UniqueGen :> r, FileSystem :> r, Concurrent :> r) => Declaration SourceRegion Shunted -> Eff r ()

Add's a declaration's name and expected type to the type environment

inferDeclarationScheme :: forall (r :: [Effect]). (Error (InferError SourceRegion) :> r, Error (UnifyError SourceRegion) :> r, Error SomeReportableError :> r, State (LocalTypeEnvironment SourceRegion) :> r, State (TypeEnvironment SourceRegion) :> r, Reader ContextStack :> r, Rock Query :> r, DiagnosticWriter (Doc AnsiStyle) :> r, UniqueGen :> r, Concurrent :> r, FileSystem :> r, StructuredDebug :> r) => Declaration SourceRegion Shunted -> Eff r (Polytype SourceRegion)

inferDeclaration :: forall (r :: [Effect]). (HasCallStack, InferPipelineEffects r, Infer SourceRegion r) => Declaration SourceRegion Shunted -> Eff r (Declaration SourceRegion Typed)

createTypeVar :: Located (Unique LowerAlphaName) -> UniqueTyVar

kindedToTypedTypeDecl :: TypeDeclaration SourceRegion Kinded -> TypeDeclaration SourceRegion Typed

Convert a kinded type declaration to a typed type declaration The only difference is the TypeVariable: Kinded uses Located (Unique LowerAlphaName), Typed uses Located UniqueTyVar. TypeMeta and ConstructorBinder are the same.

kindedToTypedType :: Type SourceRegion Kinded -> Type SourceRegion Typed

kindedToTypedType' :: Type' SourceRegion Kinded -> Type' SourceRegion Typed

inferAnnotation :: forall (r :: [Effect]). (Error (UnifyError SourceRegion) :> r, Error (InferError SourceRegion) :> r, Error SomeReportableError :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Rock Query :> r, State (TypeEnvironment SourceRegion) :> r, State (LocalTypeEnvironment SourceRegion) :> r, Reader ContextStack :> r, StructuredDebug :> r, UniqueGen :> r, Concurrent :> r, FileSystem :> r) => Annotation SourceRegion Shunted -> Eff r (Annotation SourceRegion Typed)

inferAnnotationArg :: forall (r :: [Effect]). (Error (UnifyError SourceRegion) :> r, Error (InferError SourceRegion) :> r, Error SomeReportableError :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Rock Query :> r, State (TypeEnvironment SourceRegion) :> r, State (LocalTypeEnvironment SourceRegion) :> r, Reader ContextStack :> r, StructuredDebug :> r, UniqueGen :> r, Concurrent :> r, FileSystem :> r) => AnnotationArg SourceRegion Shunted -> Eff r (AnnotationArg SourceRegion Typed)

inferValue :: forall (r :: [Effect]). (Error (UnifyError SourceRegion) :> r, Reader ContextStack :> r, State (TypeEnvironment SourceRegion) :> r, State (LocalTypeEnvironment SourceRegion) :> r, Rock Query :> r, Error (InferError SourceRegion) :> r, Error SomeReportableError :> r, DiagnosticWriter (Doc AnsiStyle) :> r, UniqueGen :> r, FileSystem :> r, Concurrent :> r, StructuredDebug :> r) => Qualified VarName -> ShuntedExpr -> Maybe (Type SourceRegion) -> Eff r (TypedExpr, Polytype SourceRegion)

exprLocation :: forall {k} loc (p :: k). Expr loc p -> loc

Get the location of an expression

skolemise :: forall (r :: [Effect]). Type SourceRegion -> Eff r (Monotype SourceRegion)

newtype SubstitutableExpr (loc :: k)

Constructors

SubstitutableExpr

Fields

getExpr :: TypedExpr

substituteExpr' :: UniqueTyVar -> Monotype SourceRegion -> TypedExpr' -> TypedExpr'

Substitute a type variable in an expression body

substituteAllExpr' :: Substitution SourceRegion -> TypedExpr' -> TypedExpr'

Bulk substitute in expression body (performance optimized)

substituteExpr :: UniqueTyVar -> Monotype SourceRegion -> TypedExpr -> TypedExpr

substituteAllExpr :: Substitution SourceRegion -> TypedExpr -> TypedExpr

substitutePattern :: UniqueTyVar -> Monotype SourceRegion -> TypedPattern -> TypedPattern

substituteAllPattern :: Substitution SourceRegion -> TypedPattern -> TypedPattern

substitutePattern' :: UniqueTyVar -> Monotype SourceRegion -> TypedPattern' -> TypedPattern'

substituteAllPattern' :: Substitution SourceRegion -> TypedPattern' -> TypedPattern'

substituteLambdaBinder :: UniqueTyVar -> Monotype SourceRegion -> TypedLambdaParam (Unique VarName) SourceRegion Typed -> TypedLambdaParam (Unique VarName) SourceRegion Typed

substituteAllLambdaBinder :: Substitution SourceRegion -> TypedLambdaParam (Unique VarName) SourceRegion Typed -> TypedLambdaParam (Unique VarName) SourceRegion Typed

substituteAstType :: UniqueTyVar -> Monotype SourceRegion -> TypedType -> TypedType

Substitute a type variable in an AST type (New.Type SourceRegion Typed)

substituteAstType' :: UniqueTyVar -> Monotype SourceRegion -> TypedType' -> TypedType'

substituteAllAstType :: Substitution SourceRegion -> TypedType -> TypedType

Bulk substitute in an AST type

monotypeToAstType :: SourceRegion -> ElaraKind -> Monotype SourceRegion -> TypedType

Convert a Monotype to an AST Type for the Typed phase

typeVarToUniqueTyVar :: TypeVariable -> UniqueTyVar

Orphan instances

PhaseCoerce (Exposing loc Shunted) (Exposing loc Typed)

Instance details

Methods

phaseCoerce :: Exposing loc Shunted -> Exposing loc Typed

PhaseCoerce (Exposition loc Shunted) (Exposition loc Typed)

Instance details

Methods

phaseCoerce :: Exposition loc Shunted -> Exposition loc Typed

PhaseCoerce (Import loc Shunted) (Import loc Typed)

Instance details

Methods

phaseCoerce :: Import loc Shunted -> Import loc Typed

PhaseCoerce (Import' loc Shunted) (Import' loc Typed)

Instance details

Methods

phaseCoerce :: Import' loc Shunted -> Import' loc Typed

PhaseCoerce (ImportExposingOrHiding loc Shunted) (ImportExposingOrHiding loc Typed)

Instance details

Methods

phaseCoerce :: ImportExposingOrHiding loc Shunted -> ImportExposingOrHiding loc Typed