Elara.TypeInfer.ConstraintGeneration

Documentation

data UnifyError loc

Constructors

OccursCheckFailed TypeVariable (Monotype loc)

TypeConstructorMismatch (loc, Qualified TypeName) (loc, Qualified TypeName)

ArityMismatch

UnificationFailed (Maybe (Constraint loc)) (Monotype loc, Monotype loc)

UnifyMismatch

UnresolvedConstraint (Qualified VarName) (Constraint loc)

PatternConstructorArityMismatch DataCon Int Int loc

PolytypeAlias ([UniqueTyVar], Type loc)

generateConstraints :: forall (r :: [Effect]) loc. ConstraintGenEffects r loc => ShuntedExpr -> Eff r (TypedExpr, Monotype loc)

solveConstraint :: forall (r :: [Effect]) a. (UnifyEffects r a, Monoid (Constraint a)) => Constraint a -> Set UniqueTyVar -> Constraint a -> Eff r (Constraint a, Substitution a)

lookupType :: forall (r :: [Effect]) loc. (StructuredDebug :> r, QueryEffects r, loc ~ SourceRegion, Rock Query :> r, State (TypeEnvironment loc) :> r) => TypeEnvKey loc -> Eff r (Type loc)

Lookup a type in the type environment, querying if necessary. Looks in the local environment first to avoid unnecessary queries / infinite loops, then queries the global environment if not found locally.