Elara.TypeInfer.Context

Description

Inference context tracking for better error messages

Documentation

data InferenceContext

Describes the context in which type inference is happening. This is used to provide better error messages by explaining WHY two types are being compared.

Constructors

CheckingFunctionArgument
Checking an argument to a function call

Fields

argPosition :: !Int

functionName :: !(Maybe (Qualified VarName))

callSite :: !SourceRegion

CheckingFunctionResult
Checking the result type of a function call

Fields

callSite :: !SourceRegion

CheckingIfCondition
Checking the condition of an if expression

Fields

ifSite :: !SourceRegion

CheckingIfBranches
Checking that if branches have the same type

Fields

thenSite :: !SourceRegion

elseSite :: !SourceRegion

CheckingMatchBranch
Checking a branch in a match expression

Fields

branchIndex :: !Int

branchSite :: !SourceRegion

CheckingPattern
Checking a pattern

Fields

patternSite :: !SourceRegion

CheckingLetBinding
Checking a let binding

Fields

bindingName :: !(Qualified VarName)

bindingSite :: !SourceRegion

CheckingAnnotation
Checking against a type annotation

Fields

annotationSite :: !SourceRegion

newtype ContextStack

A stack of inference contexts, with the most recent context at the head.

Constructors

ContextStack [InferenceContext]

emptyContextStack :: ContextStack

Create an empty context stack

pushContext :: InferenceContext -> ContextStack -> ContextStack

Push a context onto the stack

currentContext :: ContextStack -> Maybe InferenceContext

Get the current (most recent) context, if any

allContexts :: ContextStack -> [InferenceContext]

Get all contexts in the stack (most recent first)