Elara.TypeInfer.Error

Description

Type inference error types with rich context for error reporting.

Error Types

data UnifyError loc

Enriched unification error with full context for error reporting

Constructors

UnifyError

Fields

ueKind :: !UnifyErrorKind

What kind of error occurred

ueExpected :: !(Monotype loc)

The expected type

ueActual :: !(Monotype loc)

The actual type found

ueExpectedUsage :: !loc

Where the expected type was inferred/used

ueActualUsage :: !loc

Where the actual type was found/provided

ueConstraintSite :: !loc

Where the constraint was generated

ueContext :: !ContextStack

Why we were comparing these types

UnresolvedConstraint (Qualified VarName) (Constraint loc)
Unresolved constraint at the end of type checking

PolytypeAlias ([UniqueTyVar], Type loc)
Polytypes can't be used as type aliases (legacy, keep for backwards compat)

data UnifyErrorKind

The kind of unification error that occurred

Constructors

TypeMismatch
Two types couldn't be unified

TypeConstructorMismatch (Qualified TypeName) (Qualified TypeName)
Two type constructors don't match

OccursCheck TypeVariable
Occurs check failed (infinite type)

ArityMismatch Int Int
Arity mismatch in type constructor application

PatternArityMismatch DataCon Int Int
Pattern constructor arity mismatch

UnifyMismatch
Internal unification failed

PolytypeAliasError
Polytypes can't be used as type aliases

Error Construction

mkUnifyError :: UnifyErrorKind -> Monotype loc -> Monotype loc -> loc -> ContextStack -> UnifyError loc

Create a UnifyError from a constraint and error kind

mkUnifyErrorFromConstraint :: UnifyErrorKind -> Monotype loc -> Monotype loc -> Constraint loc -> ContextStack -> UnifyError loc

Create a UnifyError from a constraint, using its context information