Elara.TypeInfer.Type

Description

Types used by the type inference engine

Type Variables

data TypeVariable

Constructors

UnificationVar UniqueTyVar

SkolemVar UniqueTyVar

type UniqueTyVar = Unique (Maybe Text)

Types

data Type loc

A type scheme σ

Constructors

Polytype !(Polytype loc)

Lifted (Monotype loc)

data Polytype loc

Constructors

Forall loc [UniqueTyVar] (Constraint loc) (Monotype loc)

data Monotype loc

A monotype τ

Constructors

TypeVar loc TypeVariable
A type variable tv

TypeConstructor loc (Qualified TypeName) [Monotype loc]
A type constructor

Function loc (Monotype loc) (Monotype loc)
A function type τ₁ → τ₂

typeLoc :: Type loc -> loc

polytypeLoc :: Polytype loc -> loc

monotypeLoc :: Monotype loc -> loc

functionMonotypeResult :: Monotype loc -> Monotype loc

functionMonotypeArgs :: Monotype loc -> [Monotype loc]

Constraints

data Constraint loc

A constraint Q

Constructors

EmptyConstraint loc
The empty constraint 𝜖

Conjunction loc (Constraint loc) (Constraint loc)
The conjunction of two constraints, Q₁ ∧ Q₂

Equality
An equality constraint, τ₁ ∼ τ₂ Now enriched with usage locations and inference context

Fields

eqLoc :: loc

Where the constraint was generated

eqLeft :: Monotype loc

The left-hand side type

eqRight :: Monotype loc

The right-hand side type

eqLeftUsage :: loc

Where the left type was used/expected

eqRightUsage :: loc

Where the right type was found/provided

eqContext :: Maybe InferenceContext

Why we're comparing these types

constraintLoc :: Constraint loc -> loc

simpleEquality :: loc -> Monotype loc -> Monotype loc -> Constraint loc

Smart constructor for simple equality constraints (no context) Uses the monotype locations as usage locations

equalityWithContext :: loc -> Monotype loc -> Monotype loc -> loc -> loc -> Maybe InferenceContext -> Constraint loc

Smart constructor for equality constraints with context

reduce :: Constraint a -> Constraint a

emptyLocation :: SourceRegion

Data Constructors

type DataCon = Qualified TypeName

Axiom Schemes

data AxiomScheme (loc :: k)

An axiom scheme QQ

Constructors

EmptyAxiomScheme
The empty axiom scheme 𝜖

ConjunctionAxiomScheme (AxiomScheme loc) (AxiomScheme loc)
The conjunction of two axiom schemes, QQ₁ ∧ QQ₂

ForallAxiomScheme
A universal quantification of an axiom scheme, ∀α. QQ ⇒ QQ Practically this could be a declaration like forall a. Eq a => Eq [a]

Fields

UniqueTyVar

Type variable to be quantified

(AxiomScheme loc)

Constraint for the quantified variable

(AxiomScheme loc)

The axiom

Substitution

newtype Substitution loc

Constructors

Substitution (Map UniqueTyVar (Monotype loc))

class Substitutable (a :: Type -> Type) loc where

Methods

substitute :: UniqueTyVar -> Monotype loc -> a loc -> a loc

substituteAll :: Substitution loc -> a loc -> a loc

Apply an entire substitution to a type, recursively until a fixed point is reached

substitution :: (UniqueTyVar, Monotype loc) -> Substitution loc