Rock

Types

type Rules (f :: [Effect] -> Type -> Type) = forall a (es :: [Effect]). (HasCallStack, StructuredDebug :> es) => f es a -> Eff es a

data Rock (f :: [Effect] -> Type -> Type) (a :: Type -> Type) b

runRock :: forall (f :: [Effect] -> Type -> Type) (es :: [(Type -> Type) -> Type -> Type]) a. Rules f -> Eff (Rock f ': es) a -> Eff es a

fetch :: forall (xs :: [Effect]) (es :: [Effect]) f a. (Subset xs es, Rock f :> es, HasCallStack, StructuredDebug :> xs) => f xs a -> Eff es a

Running tasks

Task combinators

data IOQuery (f :: [(Type -> Type) -> Type -> Type] -> k -> Type) (es :: [(Type -> Type) -> Type -> Type]) (a :: k) where

Constructors

IOQuery :: forall {k} (f :: [(Type -> Type) -> Type -> Type] -> k -> Type) (es1 :: [(Type -> Type) -> Type -> Type]) (a :: k). f es1 a -> IOQuery f (IOE ': es1) a

Utils

data HideEffects (f :: [Effect] -> Type -> Type) a where

A GADT for forgetting the effects required for each key The GEq and Eq instances will unsafeCoerce away information on the Effects, please don't rely on it.

This is used for using query keys as map keys

Constructors

HideEffects :: forall (f :: [Effect] -> Type -> Type) (b :: [Effect]) a. f b a -> HideEffects f a

Orphan instances

(forall (a :: k). Hashable (f a), GEq f) => Hashable (Some f)

Instance details

Methods

hashWithSalt :: Int -> Some f -> Int

hash :: Some f -> Int