Rock.Memo

Documentation

memoise :: forall (f :: [Effect] -> Type -> Type). (forall (es :: [Effect]). GEq (f es), forall (es :: [Effect]) a. Hashable (f es a), HasCallStack) => IORef (DHashMap (HideEffects f) MVar) -> Rules f -> Rules f

Remember what f queries have already been performed and their results in a DHashMap, and reuse them if a query is performed again a second time.

The DHashMap should typically not be reused if there has been some change that might make a query return a different result.

memoiseWithCycleDetection

Arguments

:: forall (f :: [Effect] -> Type -> Type). (Typeable f, forall (es :: [Effect]) a. Show (f es a), forall (es :: [Effect]). GEq (f es), forall (es :: [Effect]) a. Hashable (f es a), forall (es :: [Effect]) a. Show (f es a), HasCallStack)

=> IORef (DHashMap (HideEffects f) MemoEntry)
started queries

-> IORef (HashMap ThreadId ThreadId)
dependencies between threads

-> Rules f

-> Rules f

Like memoise, but throw Cyclic f if a query depends on itself, directly or indirectly.

The HashMap represents dependencies between threads and should not be reused between invocations.

withoutMemoisation :: forall (f :: [Effect] -> Type -> Type). Rules f -> Rules (MemoQuery f)

memoiseExplicit :: forall (f :: [Effect] -> Type -> Type). (forall (es :: [Effect]). GEq (f es), forall (es :: [Effect]) a. Hashable (f es a)) => IORef (DHashMap (HideEffects f) MVar) -> Rules f -> Rules (MemoQuery f)

Remember what f queries have already been performed and their results in a DHashMap, and reuse them if a query is performed again a second time.

The DHashMap should typically not be reused if there has been some change that might make a query return a different result.