Rock.MemoE

Documentation

data Memoise (a :: Type -> Type) b where

a memoisation effect that's basically just a very restricted IORef This means that queries can depend on this effect without being able to bring in everything in IO

Constructors

GetStartedVar :: forall (f :: [Effect] -> Type -> Type) (a :: Type -> Type). (forall (es :: [Effect]). GEq (f es), forall (es :: [Effect]) a1. Hashable (f es a1)) => Memoise a (DHashMap (HideEffects f) MVar')

ModifyStartedVar :: forall (f :: [Effect] -> Type -> Type) (a :: Type -> Type) b. (DHashMap (HideEffects f) MVar' -> (DHashMap (HideEffects f) MVar', b)) -> Memoise a b

modifyStartedVar :: forall (f :: [Effect] -> Type -> Type) b (es :: [Effect]). (HasCallStack, Memoise :> es) => (DHashMap (HideEffects f) MVar' -> (DHashMap (HideEffects f) MVar', b)) -> Eff es b

Perform the operation ModifyStartedVar.

getStartedVar :: forall (f :: [Effect] -> Type -> Type) (es :: [Effect]). (HasCallStack, Memoise :> es, forall (es1 :: [Effect]). GEq (f es1), forall (es1 :: [Effect]) a. Hashable (f es1 a)) => Eff es (DHashMap (HideEffects f) MVar')

Perform the operation GetStartedVar.

memoiseAsIO :: forall (es :: [Effect]) a. IOE :> es => (forall (f :: [Effect] -> Type -> Type). IORef (DHashMap (HideEffects f) MVar')) -> Eff (Memoise ': es) a -> Eff es a

memoiseRunIO :: forall (f :: [Effect] -> Type -> Type) (es :: [Effect]) b. (forall (es1 :: [Effect]). GEq (f es1), forall (es1 :: [Effect]) a. Hashable (f es1 a), IOE :> es) => Eff (Memoise ': es) b -> Eff es b