Elara.Core.TypeCheck

Description

Type *checking* for the Core language. This functions as a sanity check to make sure that any optimisations haven't broken the structure of the program to form an invalid program

It doesn't do any inference! As Core is already typed, it just checks that the types are consistent

Documentation

typeCheckCoreModule :: forall (r :: [Effect]). (Error TypeCheckError :> r, HasCallStack, StructuredDebug :> r) => CoreModule (Bind Var Expr) -> Eff r ()

data TypeCheckError

Constructors

UnknownVariable

Fields

unknownVariable :: Var

unknownVariableScope :: Set (UnlocatedVarRef Text)

CoreTypeMismatch

Fields

expected :: Type

actual :: Type

source :: (CoreExpr, CoreExpr)

errorCallStack :: CallStack

CoreTypeMismatchIncompleteExpected

Fields

incompleteExpected :: Text

actual :: Type

source :: (CoreExpr, CoreExpr)

UnificationError CoreExpr CoreExpr

InfiniteType Var CoreExpr

OccursCheck Var CoreExpr

PatternMatchMissingBinders

Fields

alt :: AltCon

altType :: Type

providedBinders :: [Var]

expr :: CoreExpr