Elara.Data.Kind.Infer

Description

Performs kind inference on a type. Algorithm inspired by https://gilmi.me/blog/post/2023/09/30/kind-inference

> infer "type alias Identity a = a"

"k -> k"

> infer "type Maybe a = Either a Unit"

Type -> Type

> infer "type Either a b = Left a | Right b"

Type -> Type -> Type

> infer "type Proxy a = Proxy"

k -> Type

TODO: this is extremely basic and needs some nicer error messages, no support for polykinds or higher-kinded types yet

Documentation

data InferState

Constructors

InferState

Fields

env :: Map (Either (Qualified TypeName) TypeVar) KindVar

a mapping from type variables and named types to their kind variables

kindEnv :: Map (Qualified TypeName) ElaraKind

predefined kinds for named types

constraints :: [(ElaraKind, ElaraKind)]

substitution :: Map KindVar ElaraKind

type KindInfer (r :: [Effect]) = (State InferState :> r, Error KindInferError :> r, UniqueGen :> r, HasCallStack, QueryEffects r, Rock Query :> r)

initialInferState :: InferState

data KindInferError

Constructors

UnknownKind (Qualified TypeName) (Map (Qualified TypeName) ElaraKind)

UnboundVar TypeVar (Map (Either (Qualified TypeName) TypeVar) KindVar)

CannotUnify ElaraKind ElaraKind

NotFunctionKind ElaraKind

OccursCheckFailed KindVar ElaraKind

declareTypeVar :: forall (r :: [Effect]). State InferState :> r => TypeVar -> KindVar -> Eff r ()

declareNamedType :: forall (r :: [Effect]). State InferState :> r => Qualified TypeName -> KindVar -> Eff r ()

elaborate :: forall (r :: [Effect]). KindInfer r => Qualified TypeName -> [Located (Unique LowerAlphaName)] -> ShuntedTypeDeclaration -> Eff r (KindVar, MidKindedTypeDeclaration)

infer :: forall (r :: [Effect]). KindInfer r => Map (Qualified TypeName) ElaraKind -> [(Qualified TypeName, [Located TypeVar], ShuntedTypeDeclaration)] -> Eff r [(ElaraKind, KindedTypeDeclaration)]

inferKind :: forall (r :: [Effect]). (State InferState :> r, Error KindInferError :> r, Error SomeReportableError :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Rock Query :> r, UniqueGen :> r, FileSystem :> r, Memoise :> r, Concurrent :> r, StructuredDebug :> r) => Qualified TypeName -> [Located TypeVar] -> ShuntedTypeDeclaration -> Eff r (ElaraKind, KindedTypeDeclaration)

inferTypeKind :: forall (r :: [Effect]). KindInfer r => ShuntedType -> Eff r KindedType

lookupKindVarInSubstitution :: forall (r :: [Effect]). KindInfer r => KindVar -> Eff r ElaraKind

lookupNameKindVar :: forall (r :: [Effect]). (State InferState :> r, UniqueGen :> r, Error KindInferError :> r, QueryEffects r, Rock Query :> r) => Qualified TypeName -> Eff r KindVar

lookupVarKindVar :: forall (r :: [Effect]). KindInfer r => TypeVar -> Eff r KindVar

Find the kind variable of a type variable in the environment.

lookupKindVarMaybe :: forall (r :: [Effect]). KindInfer r => Either (Qualified TypeName) TypeVar -> Eff r (Maybe KindVar)

newEqualityConstraint :: forall (r :: [Effect]). State InferState :> r => ElaraKind -> ElaraKind -> Eff r ()

elaborateType :: forall (r :: [Effect]). KindInfer r => ShuntedType -> Eff r MidKindedType

solveType :: forall (r :: [Effect]). KindInfer r => MidKindedType -> Eff r KindedType

getAnn :: KindedType -> ElaraKind

getAnn' :: MidKindedType -> ElaraKind

solveConstraints :: forall (r :: [Effect]). (State InferState :> r, Error KindInferError :> r, HasCallStack, UniqueGen :> r, QueryEffects r, Rock Query :> r) => Eff r ()

instantiate :: forall (r :: [Effect]). UniqueGen :> r => ElaraKind -> [KindVar] -> Eff r ElaraKind

replaceKindVarWithFreshKindVar :: forall (r :: [Effect]). UniqueGen :> r => ElaraKind -> KindVar -> Eff r ElaraKind

replaceInState :: forall (r :: [Effect]). KindInfer r => KindVar -> ElaraKind -> Eff r ()

occursCheck :: forall (r :: [Effect]). KindInfer r => KindVar -> ElaraKind -> Eff r ()