Elara.Desugar

Documentation

type Desugar a = Eff DesugarPipelineEffects a

type DesugarPipelineEffects = '[State DesugarState, Error DesugarError]

newtype DesugarState

Constructors

DesugarState

Fields

_partialDeclarations :: Map (IgnoreLocation Name) PartialDeclaration

partialDeclarations :: Iso' DesugarState (Map (IgnoreLocation Name) PartialDeclaration)

resolvePartialDeclaration :: PartialDeclaration -> Desugar DesugaredDeclarationBody

resolveAnn :: Maybe (ValueDeclAnnotations 'Desugared) -> ValueDeclAnnotations 'Desugared

getDesugaredModule :: ModuleName -> Eff (ConsQueryEffects '[Error DesugarError, Rock Query]) (Module 'Desugared)

desugar :: Module 'Frontend -> Desugar (Module 'Desugared)

desugarDeclarations :: Located ModuleName -> [FrontendDeclaration] -> Desugar [DesugaredDeclaration]

assertPartialNamesEqual :: Eq a => (PartialDeclaration, Located a) -> (PartialDeclaration, Located a) -> Desugar ()

mergeAnnotations :: Maybe (ValueDeclAnnotations 'Desugared) -> Maybe (ValueDeclAnnotations 'Desugared) -> Desugar (ValueDeclAnnotations 'Desugared)

mergePartials :: PartialDeclaration -> PartialDeclaration -> Desugar PartialDeclaration

genPartials :: [FrontendDeclaration] -> Desugar ()

desugarAnnotation :: Annotation 'Frontend -> Desugar (Annotation 'Desugared)

desugarAnnotationArg :: AnnotationArg 'Frontend -> Desugar (AnnotationArg 'Desugared)

desugarType :: FrontendType' -> Desugar DesugaredType'

completePartials :: Located ModuleName -> Desugar [DesugaredDeclaration]

desugarExpr :: FrontendExpr -> Desugar DesugaredExpr

desugarBinaryOperator :: FrontendBinaryOperator -> Desugar DesugaredBinaryOperator

desugarPattern :: FrontendPattern -> Desugar DesugaredPattern

foldLambda :: [DesugaredPattern] -> DesugaredExpr -> DesugaredExpr

Lambdas need quite a lot of desugaring - they need to be unfolded into a series of nested lambdas, and then each pattern needs to be converted into a match expression.

For example, a (b, c) 1 -> e becomes a -> b_1 -> c_1 -> match b_1 with (b, c) -> match 1 with 1 -> e

However, converting the matches would require renaming, and we're not able to do that yet. Instead, we unfold the lambda, but keep the patterns, and the renamer handles the match conversion.