Elara.Error.Codes

Documentation

type ErrorCode = forall ann. Doc ann

A type for error codes. Error codes should follow a standard format consisting of the following components:

[E|W], followed by 4 digits. The symbol E indicates an error, while W indicates a warning.

The digits are used to identify the error or warning.

The first digit signals the stage in which the error was thrown:

0: Lexer / Parser

1: Desugaring and Renaming stage

2: Operator Shunting Stage

3: Typechecking Stage

4: Code Generation Stage

The remaining digits are arbitrary and should be incremented for each new error or warning. They should be unique overall, but don't have to be unique within a stage or category. For example E1001, E0001, and W1001 can all exist at once.

Some errors are internal, rather than user-facing (i.e. something internal has gone wrong if they are thrown). These errors' codes should be prefixed with [Internal].

fileReadError :: ErrorCode

fileNotFound :: ErrorCode

tooMuchIndentation :: ErrorCode

genericLexicalError :: ErrorCode

unterminatedStringLiteral :: ErrorCode

genericParseError :: ErrorCode

unknownName :: ErrorCode

unknownModule :: ErrorCode

ambiguousName :: ErrorCode

qualifiedWithWrongModule :: ErrorCode

duplicateDefinition :: ErrorCode

defWithoutLet :: ErrorCode

partialNamesNotEqual :: ErrorCode

tooManyDeclarations :: ErrorCode

nonExistentModuleDeclaration :: ErrorCode

infixDeclarationWithoutValue :: ErrorCode

duplicateFixityAnnotations :: ErrorCode

tuplePatternTooShort :: ErrorCode

samePrecedence :: ErrorCode

unknownPrecedence :: ErrorCode

unknownVariableTC :: ErrorCode

coreTypeMismatch :: ErrorCode

invokeStaticLocal :: ErrorCode

localVariableNotFound :: ErrorCode