Elara.JVM.IR

Description

Module for JVM Intermediate Representation.

Documentation

data Module

Constructors

Module

Fields

moduleName :: QualifiedClassName

The source module name

moduleClasses :: [Class]

All classes generated from this module

data Class

Constructors

Class

Fields

className :: QualifiedClassName

classSuper :: QualifiedClassName

classFields :: [Field]

classMethods :: [Method]

classConstructors :: [Constructor]

data Field

Constructors

Field

Fields

fieldName :: Text

fieldType :: FieldType

data Method

Constructors

Method

Fields

methodName :: Text

methodDesc :: MethodDescriptor

methodArgs :: [(Unique Text, FieldType)]

Argument names and types

methodBody :: [Block]

Method body as a list of basic blocks

methodIsStatic :: Bool

data Constructor

A class's constructor

Constructors

Constructor

Fields

constructorDesc :: MethodDescriptor

constructorArgs :: [(Unique Text, FieldType)]

constructorBody :: [Block]

data Block

A basic block of JVM instructions

Constructors

Block

Fields

blockLabel :: Unique Text

instrs :: [Instruction]

data Instruction

Abstract representation of JVM instructions

Constructors

Assign (Unique Text) FieldType Expr
r1 = ...

ExprStmt Expr
execute expr, drop result

Return (Maybe Expr)
return x; (for object types, uses AReturn)

IReturn Expr
return primitive int/boolean; (uses IReturn)

Jump (Unique Text)
goto label

JumpIf Expr (Unique Text) (Unique Text)
if (cond) goto true else goto false (cond is Elara Bool)

JumpIfPrimitiveBool Expr (Unique Text) (Unique Text)
if (primBool) goto true else goto false (primBool is primitive boolean)

SetField QualifiedClassName Text FieldType Expr
Set a field on this (implicit this access) e.g. "this.x = y"

Super QualifiedClassName [Expr]
Call the superclass constructor (super(...)) The Emitter will resolve the correct "invokespecial" logic.

data Expr

Constructors

LitInt Integer
Literal integer

LitString Text
Literal string

LitDouble Double
Literal double

LitBool Bool
Literal boolean (produces ElaraPrimBool)

PrimitiveLitBool Bool
Primitive boolean literal (produces JVM primitive boolean)

LitChar Char
Literal char

LitUnit

LocalVar (Unique Text) FieldType
Local Variable reference, the variable guaranteed to be in scope

This FieldType
Reference to this in instance methods

FieldRef QualifiedClassName Text FieldType

New QualifiedClassName [(Expr, FieldType)]
Create new object with constructor args

GetField
obj.field

Fields

Expr

ClassInfoType

Name of the field's class

Text

FieldType

BinaryOp BinOp Expr Expr
a + b

UnaryOp Op Expr
!a

InstanceOf Expr FieldType
obj instanceof Type

Call CallType [Expr]
method call with arguments

MakeClosure

Fields

closureTargetClass :: QualifiedClassName

The class containing the static method

closureTargetMethod :: Text

The name of the static method to eventually invoke

closureTarget :: MethodDescriptor

The static method descriptor

closureInterface :: QualifiedClassName

The functional interface we are implementing

capturedValues :: [(Expr, FieldType)]

The arguments we have SO FAR

MakeConstructorClosure

Fields

ctorClosureClass :: QualifiedClassName

The class whose constructor we're calling

ctorClosureDesc :: MethodDescriptor

The constructor descriptor (args -> void)

ctorClosureInterface :: QualifiedClassName

The functional interface we're implementing

ctorCapturedValues :: [(Expr, FieldType)]

The arguments we have SO FAR (partial application)

PrimOp PrimOp [Expr]

Cast Expr FieldType
Cast expression to a specific type

data CallType

Constructors

InvokeStatic QualifiedClassName Text MethodDescriptor

InvokeVirtual Expr QualifiedClassName Text MethodDescriptor

InvokeInterface Expr QualifiedClassName Text MethodDescriptor

data PrimOp

Elara primitive operations

Constructors

UndefinedError

PatternMatchFailedError

IntAdd

IntSubtract

IntMultiply

IntNegate

PrimEquals

PrimCompare

IOBind

DebugWithMsg

ThrowError

Println

StringCons

StringHead

StringIsEmpty

StringTail

ToString

data BinOp

Constructors

Add

Subtract

Multiply

Divide

And

Or

Equals

NotEquals

LessThan

GreaterThan

data Op

Constructors

Not

Negate

prettyCall :: CallType -> [Expr] -> Doc AnsiStyle