Elara.Lexer.Utils

Documentation

data AlexInput

Constructors

AlexInput

Fields

_filePath :: FilePath

_prev :: Char

_bytes :: [Word8]

_rest :: Text

_position :: RealPosition

data IndentInfo

Constructors

IndentInfo

Fields

_indent :: Int

_indentPos :: RealPosition

_openedAtDepth :: Int

the delimDepth when this indent was opened

data ParseState

Constructors

ParseState

Fields

_input :: AlexInput

_lexSC :: Int

lexer start code

_stringBuf :: Text

temporary storage for strings

_pendingTokens :: [Lexeme]

right now used when Parser consumes the lookahead and decided to put it back

_indentStack :: NonEmpty IndentInfo

stack of indentation levels

_pendingPosition :: TokPosition

needed when parsing strings, chars, multi-line strings

_prevEndsExpr :: Bool

did the previous token end an expression? (used for offside rule

_delimDepth :: Int

current depth of open delimiters ((), [], {

bytes :: Lens' AlexInput [Word8]

filePath :: Lens' AlexInput FilePath

position :: Lens' AlexInput RealPosition

prev :: Lens' AlexInput Char

rest :: Lens' AlexInput Text

indent :: Lens' IndentInfo Int

indentPos :: Lens' IndentInfo RealPosition

openedAtDepth :: Lens' IndentInfo Int

delimDepth :: Lens' ParseState Int

indentStack :: Lens' ParseState (NonEmpty IndentInfo)

input :: Lens' ParseState AlexInput

lexSC :: Lens' ParseState Int

pendingPosition :: Lens' ParseState TokPosition

pendingTokens :: Lens' ParseState [Lexeme]

prevEndsExpr :: Lens' ParseState Bool

stringBuf :: Lens' ParseState Text

type LexMonad a = Eff '[State ParseState, Error LexerError, StructuredDebug] a

setPrevEndsExpr :: Bool -> LexMonad ()

getPrevEndsExpr :: LexMonad Bool

mkIndentInfo :: Int -> LexMonad IndentInfo

data LexerError

Constructors

TooMuchIndentation
When an element is indented more than expected

Fields

IndentInfo

The expected indentation

(Maybe IndentInfo)

The potential further indentation

Int

The actual indentation

ParseState

The current state of the lexer

UnterminatedStringLiteral ParseState

initialState :: FilePath -> Text -> ParseState

pushFront :: Lexeme -> LexMonad ()

pushBack :: Lexeme -> LexMonad ()

emitAt :: Token -> SourceRegion -> LexMonad (Maybe Lexeme)

popPending :: LexMonad (Maybe Lexeme)

emitCloser :: Token -> SourceRegion -> LexMonad (Maybe Lexeme)

flushLayoutBeforeCloser :: LexMonad ()

emitLayoutAt :: Token -> RealSourceRegion -> LexMonad (Maybe Lexeme)

emitIndentAt :: RealPosition -> LexMonad (Maybe Lexeme)

emitDedentAt :: TokPosition -> TokPosition -> LexMonad (Maybe Lexeme)

emitLineSepAt :: RealPosition -> LexMonad (Maybe Lexeme)

fake :: Token -> LexMonad (Maybe Lexeme)

startWhite :: Int -> Text -> LexMonad (Maybe Lexeme)

cleanIndentation :: LexMonad [Lexeme]

alexGetByte :: AlexInput -> Maybe (Word8, AlexInput)

getLineNo :: LexMonad Int

getColNo :: LexMonad Int

getPosition :: Int -> LexMonad TokPosition

createRegion :: TokPosition -> TokPosition -> LexMonad RealSourceRegion

createRegionStartingAt :: TokPosition -> LexMonad RealSourceRegion

splitQualName :: Text -> (ModuleName, Text)

splits a qualified name into the qualifier and the name. Throws an error if the name is not qualified.

Examples:

>>> splitQualName "Hello.world"
(ModuleName ("Hello" :| []),"world")

>>> splitQualName "A.B.C"
(ModuleName ("A" :| ["B"]),"C")

>>> splitQualName "Prelude..+"
(ModuleName ("Prelude" :| []),".+")

>>> splitQualName "A.!."
(ModuleName ("A" :| []),"!.")