Elara.Parse.Expression

Documentation

exprParser :: Parser FrontendExpr

locatedExpr :: Parser FrontendExpr' -> Parser FrontendExpr

letPreamble :: Parser (Located VarName, [FrontendPattern], FrontendExpr)

element :: Parser FrontendExpr

A top level element, comprised of either an expression or a let declaration Note that top level let declarations are not parsed here, but in the Elara.Parse.Declaration module

reservedWords :: Set Text

Reserved words, used to backtrack accordingly