Elara.Parse.Expression

Documentation

locatedExpr :: Parser FrontendExpr' -> Parser FrontendExpr

exprParser :: Parser FrontendExpr

element :: Parser FrontendExpr

A top level element, comprised of either an expression or a let declaration Note that top level let declarations are not parsed here, but in the Elara.Parse.Declaration module

statement :: Parser FrontendExpr

This is not a "statement" in an imperative sense, but is used to parse let expressions that are not at the top level. | The reason for having this distinction is that if let's were considered a normal expression, then something like | let x = let y = 1 would be considered valid code, which it is not. Having a separate parser means local let bindings can only be allowed | in places where they are valid, such as in the body of a function.

unannotatedExpr :: Iso' FrontendExpr (Located FrontendExpr')

binOp :: Parser (FrontendExpr -> FrontendExpr -> FrontendExpr)

cons :: Parser (FrontendExpr -> FrontendExpr -> FrontendExpr)

functionCall :: Parser (FrontendExpr -> FrontendExpr -> FrontendExpr)

operator :: Parser FrontendBinaryOperator

expression :: Parser FrontendExpr

reservedWords :: Set Text

Reserved words, used to backtrack accordingly

parensExpr :: Parser FrontendExpr

variable :: Parser FrontendExpr

constructor :: Parser FrontendExpr

unit :: Parser FrontendExpr

int :: Parser FrontendExpr

float :: Parser FrontendExpr

string :: Parser FrontendExpr

charL :: Parser FrontendExpr

match :: Parser FrontendExpr

lambda :: Parser FrontendExpr

ifElse :: Parser FrontendExpr

letPreamble :: Parser (Located VarName, [FrontendPattern], FrontendExpr)

letInExpression :: Parser FrontendExpr

letStatement :: Parser FrontendExpr

tuple :: Parser FrontendExpr

list :: Parser FrontendExpr