Elara.Prim

Description

Stores information about the primitive functions of Elara. These are still written in the source code, with a special name and value. The compiler will then replace these with the actual primitive functions.

Documentation

consName :: TypeName

cons :: MaybeQualified VarOrConName

fetchPrimitiveName :: VarName

fetchPrimitive :: VarRef VarName

stringName :: TypeName

intName :: TypeName

boolName :: TypeName

trueName :: TypeName

falseName :: TypeName

charName :: TypeName

ioName :: TypeName

fullListName :: IsString a => Qualified a

primModuleName :: ModuleName

primRegion :: SourceRegion

primLocated :: a -> Located a

mkPrimQual :: c -> Qualified c

mkPrimVarRef :: c -> Located (Qualified c)

primitiveVars :: [VarName]

primitiveTypes :: [TypeName]

primKindCheckContext :: Map (Qualified TypeName) ElaraKind

primOpTable :: OpTable