Elara.Query

Description

Documentation

type WithRock (effects :: [Effect]) = Rock Query ': effects

type family QueryEffectsOf (q :: QueryType) (ast :: LocatedAST) = (es :: [Effect]) where ...

The effects required to run a specific query | This should probably be moved to the SupportsQuery class as an associated type

Equations

QueryEffectsOf q ast = WithRock (ConsQueryEffects (QuerySpecificEffectsOf q ast))

data Query (es :: [Effect]) a where

Constructors

GetCompilerSettings :: Query (Rock Query ': MinimumQueryEffects) CompilerSettings
Query to get the compiler settings

InputFiles :: Query (Rock Query ': ConsMinimumQueryEffects '[FileSystem]) (HashSet FilePath)
Query to get all the required input files to be passed to the compiler

GetFileContents :: FilePath -> Query (Rock Query ': ConsMinimumQueryEffects '[FileSystem, Error ReadFileError, DiagnosticWriter (Doc AnsiStyle)]) FileContents
Query to get the contents of a specific file

ModulePath :: ModuleName -> Query (Rock Query ': ConsMinimumQueryEffects '[Rock Query, FileSystem]) [Char]
Query to get the file path of a module

LexedFile :: FilePath -> Query (Rock Query ': ConsQueryEffects '[Error LexerError]) [Lexeme]
Query to get the lexed tokens of a specific file

ParsedFile :: FilePath -> Query (Rock Query ': ConsQueryEffects '[Error (WParseErrorBundle TokenStream ElaraParseError)]) (Module 'Frontend)
Query to get the parsed module from a file's contents and lexed tokens

ParsedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error (WParseErrorBundle TokenStream ElaraParseError)]) (Module 'Frontend)
Query to get a parsed module by module name

DesugaredModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error DesugarError]) (Module 'Desugared)

RenamedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error RenameError]) (Module 'Renamed)

ModuleByName :: forall (ast :: LocatedAST). (Typeable ast, SupportsQuery 'QueryModuleByName ast, HasMinimumQueryEffects (QueryEffectsOf 'QueryModuleByName ast)) => ModuleName -> Query (QueryEffectsOf 'QueryModuleByName ast) (QueryReturnTypeOf 'QueryModuleByName ast)
Lookup a module by name

DeclarationByName :: forall (ast :: LocatedAST). (Typeable ast, SupportsQuery 'QueryDeclarationByName ast) => QueryArgsOf 'QueryDeclarationByName ast -> Query (QueryEffectsOf 'QueryDeclarationByName ast) (QueryReturnTypeOf 'QueryDeclarationByName ast)

RequiredDeclarationByName :: forall (ast :: LocatedAST). (Typeable ast, SupportsQuery 'QueryRequiredDeclarationByName ast) => QueryArgsOf 'QueryRequiredDeclarationByName ast -> Query (QueryEffectsOf 'QueryRequiredDeclarationByName ast) (QueryReturnTypeOf 'QueryRequiredDeclarationByName ast)

ConstructorDeclaration :: forall (ast :: LocatedAST). (Typeable ast, Show (Select 'ConRef ast), Ord (Select 'ConRef ast), Hashable (Select 'ConRef ast), SupportsQuery 'QueryConstructorDeclaration ast) => QueryArgsOf 'QueryConstructorDeclaration ast -> Query (QueryEffectsOf 'QueryConstructorDeclaration ast) (QueryReturnTypeOf 'QueryConstructorDeclaration ast)
Looks up the declaration for a data constructor

DeclarationAnnotations :: forall (ast :: LocatedAST). (Typeable ast, SupportsQuery 'DeclarationAnnotations ast) => QueryArgsOf 'QueryDeclarationByName ast -> Query (QueryEffectsOf 'DeclarationAnnotations ast) (QueryReturnTypeOf 'DeclarationAnnotations ast)

DeclarationAnnotationsOfType :: forall (ast :: LocatedAST). (Typeable ast, SupportsQuery 'DeclarationAnnotationsOfType ast) => QueryArgsOf 'DeclarationAnnotationsOfType ast -> Query (QueryEffectsOf 'DeclarationAnnotations ast) (QueryReturnTypeOf 'DeclarationAnnotations ast)

GetOpInfo :: IgnoreLocVarRef Name -> Query (Rock Query ': ConsQueryEffects '[Writer (Set ShuntWarning), Error ShuntError]) (Maybe OpInfo)

GetOpTableIn :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Map (IgnoreLocVarRef Name) OpInfo)

FreeVarsOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (HashSet (Qualified VarName))

FreeTypesOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (HashSet (Qualified TypeName))

ReachableSubgraphOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) ReachableSubgraph

GetSCCsOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) [SCC (Qualified VarName)]

SCCKeyOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) SCCKey

TypeCheckedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Module 'Typed)

TypeCheckedExpr :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Expr 'Typed)

TypeOf :: forall loc. loc ~ SourceRegion => TypeEnvKey loc -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Type loc)

InferSCC :: SCCKey -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Map (Qualified VarName) (Polytype SourceRegion))

KindOf :: Qualified TypeName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe KindVar)

GetTypeAlias
Get Information about a type alias

Fields

:: Qualified TypeName

The name of the type alias

-> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe ([UniqueTyVar], Type SourceRegion))

The type alias's type variables and body, if it exists * To Core Queries

TypeCheckedTypeDeclarations :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (HashMap (Qualified TypeName) (Declaration 'Typed))

GetCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

GetTyCon :: Qualified Text -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe TyCon)

GetDataCon :: Qualified TypeName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe DataCon)

TypeCheckedDeclaration :: Qualified Name -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Declaration 'Typed)

GetOptimisedCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

GetANFCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule (TopLevelBind Var))

GetClosureLiftedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error ClosureLiftError]) (CoreModule (TopLevelBind Var))

GetFinalisedCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

data QueryType

List of query kinds at the type level

Constructors

QueryRequiredDeclarationByName

QueryDeclarationByName

QueryModuleByName

QueryConstructorDeclaration

DeclarationAnnotations

DeclarationAnnotationsOfType

type family QueryReturnTypeOf (q :: QueryType) (ast :: a) = (r :: Type) where ...

Equations

QueryReturnTypeOf 'QueryDeclarationByName (ast :: a) = Maybe (Declaration ast)

QueryReturnTypeOf 'QueryRequiredDeclarationByName (ast :: a) = Declaration ast

QueryReturnTypeOf 'QueryModuleByName (ast :: a) = Module ast

QueryReturnTypeOf 'QueryConstructorDeclaration (ast :: a) = Declaration ast

QueryReturnTypeOf 'DeclarationAnnotations (ast :: a) = [Annotation ast]

QueryReturnTypeOf 'DeclarationAnnotationsOfType (ast :: a) = [Annotation ast]

type family QueryArgsOf (q :: QueryType) (ast :: a) where ...

Equations

QueryArgsOf 'QueryDeclarationByName (ast :: a) = Qualified Name

QueryArgsOf 'QueryRequiredDeclarationByName (ast :: a) = Qualified Name

QueryArgsOf 'QueryModuleByName (ast :: a) = ModuleName

QueryArgsOf 'QueryConstructorDeclaration (ast :: a) = Select 'ConRef ast

QueryArgsOf 'DeclarationAnnotations (ast :: a) = Qualified Name

QueryArgsOf 'DeclarationAnnotationsOfType (ast :: a) = (Qualified Name, Qualified TypeName)

class (Typeable ast, HasMinimumQueryEffects (QueryEffectsOf q ast)) => SupportsQuery (q :: QueryType) (ast :: LocatedAST) where

Associated Types

type QuerySpecificEffectsOf (q :: QueryType) (ast :: LocatedAST) :: [Effect]

Effects that are "unique" to this query. Effects included in MinimumQueryEffects should not be included here as they are expected to always be present

type QuerySpecificEffectsOf (q :: QueryType) (ast :: LocatedAST) = StandardQueryError ast

Methods

query :: QueryArgsOf q ast -> Eff (QueryEffectsOf q ast) (QueryReturnTypeOf q ast)

type family SupportsQueries (qs :: [QueryType]) (ast :: LocatedAST) = (c :: Constraint) where ...

Equations

SupportsQueries ('[] :: [QueryType]) ast = ()

SupportsQueries (q ': qs) ast = (SupportsQuery q ast, SupportsQueries qs ast)

tagQuery :: forall {w1 :: [Effect]} {w2}. Query w1 w2 -> Int

sameCtor :: HasCallStack => forall (es :: [Effect]) a b. Query es a -> Query es b -> GOrdering a b