Elara.Query

Description

Documentation

type WithRock (effects :: [Effect]) = Rock Query ': effects

Appends Rock to a list of effects. This type mainly exists to avoid a cyclic import between this module and Effects

data Query (es :: [Effect]) a where

Constructors

GetCompilerSettings :: Query (Rock Query ': MinimumQueryEffects) CompilerSettings
Query to get the compiler settings

InputFiles :: Query (Rock Query ': ConsMinimumQueryEffects '[FileSystem]) (HashSet FilePath)
Query to get all the required input files to be passed to the compiler

GetFileContents :: FilePath -> Query (Rock Query ': ConsMinimumQueryEffects '[FileSystem, DiagnosticWriter (Doc AnsiStyle)]) FileContents
Query to get the contents of a specific file

ModulePath :: ModuleName -> Query (Rock Query ': ConsMinimumQueryEffects '[Rock Query, FileSystem]) [Char]
Query to get the file path of a module

LexedFile :: FilePath -> Query (Rock Query ': ConsQueryEffects '[Error LexerError]) [Lexeme]
Query to get the lexed tokens of a specific file

ParsedFile :: FilePath -> Query (Rock Query ': ConsQueryEffects '[Error (WParseErrorBundle TokenStream ElaraParseError)]) (Module 'Frontend)
Query to get the parsed module from a file's contents and lexed tokens

ParsedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error (WParseErrorBundle TokenStream ElaraParseError)]) (Module 'Frontend)
Query to get a parsed module by module name

DesugaredModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error DesugarError]) (Module 'Desugared)

RenamedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error RenameError]) (Module 'Renamed)

ShuntedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects '[Error ShuntError]) (Module 'Shunted)

ShuntedDeclarationByName :: Qualified Name -> Query (Rock Query ': ConsQueryEffects '[Error ShuntError]) (Declaration 'Shunted)

GetOpInfo :: IgnoreLocVarRef Name -> Query (Rock Query ': ConsQueryEffects '[Writer (Set ShuntWarning)]) (Maybe OpInfo)

GetOpTableIn :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Map (IgnoreLocVarRef Name) OpInfo)

FreeVarsOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (HashSet (Qualified VarName))

ReachableSubgraphOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) ReachableSubgraph

GetSCCsOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) [SCC (Qualified VarName)]

SCCKeyOf :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) SCCKey

TypeCheckedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Module 'Typed)

TypeCheckedExpr :: Qualified VarName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Expr 'Typed)

TypeOf :: forall loc. loc ~ SourceRegion => TypeEnvKey loc -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Type loc)

InferSCC :: SCCKey -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Map (Qualified VarName) (Polytype SourceRegion))

KindOf :: Qualified TypeName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe KindVar)

GetCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

GetTyCon :: Qualified Text -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe TyCon)

GetDataCon :: Qualified TypeName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (Maybe DataCon)

GetOptimisedCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)

GetANFCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule (TopLevelBind Var))

GetClosureLiftedModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule (TopLevelBind Var))

GetFinalisedCoreModule :: ModuleName -> Query (Rock Query ': ConsQueryEffects ('[] :: [Effect])) (CoreModule CoreBind)