Elara.Rename

Description

Renaming stage of compilation This stage handles: 1. Renaming all variables, types, and type variables, adding module qualification or unique suffixes to avoid name clashes 2. Desugaring any "first-class" pattern matches into normal match expressions (eg '[] -> 1' to 'x -> match x with [] -> 1') 3. Desugaring blocks into let-in chains (and monad operations soon), eg 'let y = 1; y + 1' to 'let y = 1 in y + 1' Note that until the monad operations are implemented, we can't fully remove blocks, as we have nothing to translate 'f x; g x' into

Documentation

type RenamePipelineEffects = '[State RenameState, Error RenameError, Reader (TopologicalGraph (Module 'Desugared)), UniqueGen, StructuredDebug]

type Rename (r :: [Effect]) = (State RenameState :> r, Error RenameError :> r, UniqueGen :> r, QueryEffects r, StructuredDebug :> r, Rock Query :> r)

type InnerRename (r :: [Effect]) = (State RenameState :> r, Error RenameError :> r, UniqueGen :> r, QueryEffects r, StructuredDebug :> r, Reader (Maybe (Module 'Desugared)) :> r)

getRenamedModule :: ModuleName -> Eff (ConsQueryEffects '[Error RenameError, State RenameState, Rock Query]) (Module 'Renamed)

qualifyIn :: forall (r :: [Effect]) name. Rename r => ModuleName -> MaybeQualified name -> Eff r (Qualified name)

qualifyTypeName :: forall (r :: [Effect]). (InnerRename r, Rock Query :> r) => Located (MaybeQualified TypeName) -> Eff r (Located (Qualified TypeName))

askCurrentModule :: forall (r :: [Effect]). InnerRename r => Eff r (Module 'Desugared)

lookupGenericName :: forall (r :: [Effect]) name. (UniqueGen :> r, Rock Query :> r, Ord name, ToName name, Show name, Reader (Maybe (Module 'Desugared)) :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Error SomeReportableError :> r, Error RenameError :> r, State RenameState :> r, FileSystem :> r, StructuredDebug :> r, Memoise :> r, Concurrent :> r) => Lens' RenameState (Map name (NonEmpty (VarRef name))) -> (Located Name -> NonEmpty (VarRef name) -> RenameError) -> Located (MaybeQualified name) -> Eff r (Located (VarRef name))

lookupVarName :: forall (r :: [Effect]). (Rock Query :> r, Reader (Maybe (Module 'Desugared)) :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Error SomeReportableError :> r, Error RenameError :> r, State RenameState :> r, UniqueGen :> r, FileSystem :> r, StructuredDebug :> r, Memoise :> r, Concurrent :> r) => Located (MaybeQualified VarName) -> Eff r (Located (VarRef VarName))

lookupTypeName :: forall (r :: [Effect]). (InnerRename r, Rock Query :> r) => Located (MaybeQualified TypeName) -> Eff r (Located (Qualified TypeName))

lookupTypeVar :: forall (r :: [Effect]) {a}. (State a :> r, HasField' "typeVars" a (Map LowerAlphaName (Unique LowerAlphaName))) => LowerAlphaName -> Eff r (Maybe (Unique LowerAlphaName))

uniquify :: forall (r :: [Effect]) name. UniqueGen :> r => Located name -> Eff r (Located (Unique name))

sortDeclarations :: forall (r :: [Effect]). [RenamedDeclaration] -> Eff r [RenamedDeclaration]

Performs a topological sort of field' declarations, so as many

rename :: forall (r :: [Effect]). Rename r => Module 'Desugared -> Eff r (Module 'Renamed)

addImportsToContext :: forall (r :: [Effect]). Rename r => [Import 'Desugared] -> Eff r ()

addImportToContext :: forall (r :: [Effect]). Rename r => Import 'Desugared -> Eff r ()

getModuleFromName :: forall {r :: [Effect]}. (DiagnosticWriter (Doc AnsiStyle) :> r, Error SomeReportableError :> r, Rock Query :> r, FileSystem :> r, UniqueGen :> r, StructuredDebug :> r, Memoise :> r, Concurrent :> r) => ModuleName -> Eff r (Module 'Desugared)

addModuleToContext :: forall (r :: [Effect]). Rename r => ModuleName -> Exposing 'Desugared -> Eff r ()

addDeclarationToContext :: forall (r :: [Effect]). Rename r => Bool -> DesugaredDeclaration -> Eff r ()

ensureExistsAndExposed :: forall (r :: [Effect]). (Rock Query :> r, Reader (Maybe (Module 'Desugared)) :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Error SomeReportableError :> r, Error RenameError :> r, FileSystem :> r, UniqueGen :> r, StructuredDebug :> r, Memoise :> r, Concurrent :> r) => ModuleName -> Located Name -> Eff r ()

Ensure that a name exists in the context and is exposed

elementExistsInModule :: Module 'Desugared -> Name -> Bool

isExposingAndExists :: Module 'Desugared -> Name -> Bool

Tests that n is exposed in m I.e. that it is in the exposing list, or that the module is exposing everything

renameDeclaration :: forall (r :: [Effect]). (InnerRename r, Rock Query :> r) => DesugaredDeclaration -> Eff r RenamedDeclaration

renameAnnotation :: forall (r :: [Effect]). (InnerRename r, Reader (Maybe DesugaredDeclaration) :> r, Rock Query :> r) => Annotation 'Desugared -> Eff r (Annotation 'Renamed)

renameTypeDeclaration :: forall (r :: [Effect]). (Rock Query :> r, Reader (Maybe (Module 'Desugared)) :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Error SomeReportableError :> r, Error RenameError :> r, State RenameState :> r, StructuredDebug :> r, Concurrent :> r, Memoise :> r, FileSystem :> r, UniqueGen :> r) => ModuleName -> DesugaredTypeDeclaration -> Eff r RenamedTypeDeclaration

renameSimpleType :: forall (r :: [Effect]). (State RenameState :> r, Error RenameError :> r, Error SomeReportableError :> r, DiagnosticWriter (Doc AnsiStyle) :> r, Reader (Maybe (Module 'Desugared)) :> r, Rock Query :> r, UniqueGen :> r, FileSystem :> r, Memoise :> r, Concurrent :> r, StructuredDebug :> r) => DesugaredType -> Eff r RenamedType

renameType

Arguments

:: forall (r :: [Effect]). (InnerRename r, Rock Query :> r)

=> Bool
If new type variables are allowed - if False, this will throw an error if a type variable is not in scope This is useful for type declarations, where something like type Invalid a = b would clearly be invalid But for local type annotations, we want to allow this, as it may be valid

-> DesugaredType'

-> Eff r RenamedType'

Renames a type, qualifying type constructors and type variables where necessary

renameExpr :: forall (r :: [Effect]). (InnerRename r, Reader (Maybe DesugaredDeclaration) :> r, Rock Query :> r) => DesugaredExpr -> Eff r RenamedExpr

renameBinaryOperator :: forall (r :: [Effect]). (InnerRename r, Rock Query :> r) => DesugaredBinaryOperator -> Eff r RenamedBinaryOperator

renamePattern :: forall (r :: [Effect]). (InnerRename r, Rock Query :> r) => DesugaredPattern -> Eff r RenamedPattern

patternToVarName :: DesugaredPattern -> VarName

Estimates a var name from a pattern This isn't really necessary as names will be uniquified anyway, but it could make dumped code more readable

patternToMatch :: forall (r :: [Effect]). (InnerRename r, Reader (Maybe DesugaredDeclaration) :> r, Rock Query :> r) => DesugaredPattern -> DesugaredExpr -> Eff r (Located (Unique VarName), RenamedExpr)

renameLambda :: forall (r :: [Effect]). (InnerRename r, Reader (Maybe DesugaredDeclaration) :> r, Rock Query :> r) => DesugaredPattern -> DesugaredExpr -> Eff r RenamedExpr'

Rename a lambda expression This is a little bit special because patterns have to be converted to match expressions

For example, (a, b) -> a becomes ab_ -> match ab_ with (a, b) -> a

desugarBlock :: forall (r :: [Effect]). (InnerRename r, Reader (Maybe DesugaredDeclaration) :> r, Rock Query :> r) => NonEmpty DesugaredExpr -> Eff r RenamedExpr

Orphan instances

SupportsQuery 'QueryModuleByName 'Renamed

Instance details

Associated Types

type QuerySpecificEffectsOf 'QueryModuleByName 'Renamed

Instance details

Defined in Elara.Rename

type QuerySpecificEffectsOf 'QueryModuleByName 'Renamed = StandardQueryError 'Renamed

Methods

query :: QueryArgsOf 'QueryModuleByName 'Renamed -> Eff (QueryEffectsOf 'QueryModuleByName 'Renamed) (QueryReturnTypeOf 'QueryModuleByName 'Renamed)