Elara.Rename

Description

Renaming stage of compilation This stage handles: 1. Renaming all variables, types, and type variables, adding module qualification or unique suffixes to avoid name clashes 2. Desugaring any "first-class" pattern matches into normal match expressions (eg '[] -> 1' to 'x -> match x with [] -> 1') 3. Desugaring blocks into let-in chains (and monad operations soon), eg 'let y = 1; y + 1' to 'let y = 1 in y + 1' Note that until the monad operations are implemented, we can't fully remove blocks, as we have nothing to translate 'f x; g x' into

Documentation

getRenamedModule :: ModuleName -> Eff (ConsQueryEffects '[Error RenameError, State RenameState, Rock Query]) (Module 'Renamed)

Run the QueryModuleByName Renamed query, renaming a module by its name

Orphan instances

SupportsQuery 'QueryModuleByName 'Renamed

Instance details

Associated Types

type QuerySpecificEffectsOf 'QueryModuleByName 'Renamed

Instance details

Defined in Elara.Rename

type QuerySpecificEffectsOf 'QueryModuleByName 'Renamed = StandardQueryError 'Renamed

Methods

query :: QueryArgsOf 'QueryModuleByName 'Renamed -> Eff (QueryEffectsOf 'QueryModuleByName 'Renamed) (QueryReturnTypeOf 'QueryModuleByName 'Renamed)