Elara.Rename.Imports

Description

This module mostly just exists to be a circuit breaker :)

Documentation

isImportedBy :: Module 'Desugared -> VarRef Name -> Bool

Tests that m imports n This is determined by 2 conditions: n's module name is in the import list of m n is exposed by the import