Elara.ToCore.Match

Description

Pattern Match Compilation (Maranget's Algorithm)

This module compiles high-level pattern matching into low-level Core match expressions which only contain a single branch at once.

The Pattern Matrix

We treat the match expressions as a matrix. Given (in pseudo-Elara): match x, y with Just a, Right b -> 1 _, Left c -> 2

The initial matrix is constructed as: Col 0 Col 1 Row 1: [ Just a ] [ Right b ] => Body 1 Row 2: [ _ ] [ Left c ] => Body 2 with scrutinees: [x, y]

The Algorithm Step

We look at **Column 0** (the head) to decide what to do with variable x. 1. We see a constructor Just. We must generate a match x with .... 2. We create a branch for Just. Inside this branch, x is unwrapped into payload. We transform the matrix for this branch (Specialization): - Row 1 matches Just a. The pattern Just a becomes a. - Row 2 is a wildcard _. It matches Just too! We expand _ into wildcard_payload.

New Matrix for Just branch: New Col Old Col 1 Row 1: [ a ] [ Right b ] Row 2: [ _ ] [ Left c ] New Scrutinees: [payload, y]

We recurse until the matrix is empty.

Documentation

buildMatrix1 :: [(TypedPattern, a)] -> PMatrix a

Initialise a matrix from a list of match arms. This matrix typically starts with a single column (the top-level pattern matches). As we decompose constructors, the matrix will grow wider.

compileMatrix

Arguments

:: forall (r :: [Effect]). StructuredDebug :> r

=> ConResolver (Eff r)
Function to resolve constructor names

-> FreshLocal (Eff r)
Function to generate fresh local variables

-> [Var]
The scrutinee variables

-> PMatrix CoreExpr
The pattern matrix

-> Eff r CoreExpr
The compiled Core expression

Compile a pattern matrix to a Core expression.

Assumes the first column of the matrix corresponds to the first variable in the scruts list. This function is recursive: 1. If the matrix is empty (width 0), we are done, so we emit the RHS of the first valid row. 2. If we have patterns but no scrutinees, we emit a unit value (which indicates a non-exhaustive match). 3. Otherwise, we inspect the first column (processHeadColumn) to decide how to branch, and recurse.