Elara.TypeInfer.ConstraintGeneration

Documentation

generateConstraints :: forall loc (r :: [Effect]). (Infer loc r, Error (UnifyError loc) :> r, loc ~ SourceRegion) => ShuntedExpr -> Eff r (TypedExpr, Monotype loc)

lookupType :: forall (r :: [Effect]) loc. (QueryEffects r, loc ~ SourceRegion, Rock Query :> r, State (TypeEnvironment loc) :> r) => TypeEnvKey loc -> Eff r (Type loc)

generateConstraints' :: forall loc (r :: [Effect]). (Infer loc r, Error (UnifyError loc) :> r, loc ~ SourceRegion) => ShuntedExpr -> Eff r (TypedExpr', Monotype loc)

generatePatternConstraints :: forall loc (r :: [Effect]). (Infer loc r, Error (UnifyError loc) :> r, loc ~ SourceRegion) => ShuntedPattern -> Monotype loc -> Eff r (TypedPattern, Monotype loc)

Generate constraints for a pattern, returning the typed pattern and the type of the pattern, and generating constraints for the pattern We define the type of a pattern as the type of values it can match against, rather than its "signature"

For example, in the case of a simple option type type Option a = Some a | None, we say that the pattern `Some x` has type `Option a` rather than a -> Option a

generatePatternConstraints'

Arguments

:: forall loc (r :: [Effect]). (Infer loc r, Error (UnifyError loc) :> r, loc ~ SourceRegion)

=> ShuntedPattern'

-> Monotype loc
the type of the pattern we are matching against For example if we have `match (x : Option Int) with { Some y -> y }` then over would be `Option Int` This is necessary for WildcardPattern and VarPattern to know what type they should be

-> Eff r (TypedPattern', Monotype loc)

instantiate :: forall (r :: [Effect]) loc. (Pretty loc, Eq loc, Monoid (Constraint loc), Infer loc r) => Type loc -> Eff r (Monotype loc, [TypeVariable])

solveConstraint :: forall a (r :: [Effect]). (Error (UnifyError a) :> r, StructuredDebug :> r, Pretty a, Eq a, Show a, HasCallStack, Monoid (Constraint a), Semigroup a) => Constraint a -> Set UniqueTyVar -> Constraint a -> Eff r (Constraint a, Substitution a)

simplifyConstraint :: forall loc (r :: [Effect]). (Error (UnifyError loc) :> r, StructuredDebug :> r, Pretty loc, Eq loc, Show loc, HasCallStack, Monoid (Constraint loc), Semigroup loc) => Constraint loc -> Set UniqueTyVar -> Constraint loc -> Eff r (Constraint loc, Substitution loc)

reduceGiven :: forall loc (r :: [Effect]). (Pretty loc, Error (UnifyError loc) :> r, StructuredDebug :> r, Eq loc, Show loc, HasCallStack) => Constraint loc -> Eff r (Substitution loc)

solve :: forall loc (r :: [Effect]). (Error (UnifyError loc) :> r, StructuredDebug :> r, Reader (Set UniqueTyVar) :> r, Eq loc, Show loc, HasCallStack, Monoid (Constraint loc), Semigroup loc, Pretty loc) => Constraint loc -> Eff r (Constraint loc, Substitution loc)

unifyGiven :: forall loc (r :: [Effect]). (Pretty loc, Error (UnifyError loc) :> r, StructuredDebug :> r, Show loc, Eq loc) => Maybe (Constraint loc) -> Monotype loc -> Monotype loc -> Eff r (Substitution loc)

unifyGivenMany :: forall {loc} {r :: [Effect]}. (Pretty loc, Error (UnifyError loc) :> r, StructuredDebug :> r, Show loc, Eq loc) => Maybe (Constraint loc) -> [Monotype loc] -> [Monotype loc] -> Eff r (Substitution loc)

unify :: forall loc (r :: [Effect]). (HasCallStack, StructuredDebug :> r, Error (UnifyError loc) :> r, Reader (Set UniqueTyVar) :> r, Pretty loc, Pretty (Constraint loc), Eq loc, Show loc, ?constraint :: Maybe (Constraint loc), Monoid (Constraint loc), Semigroup loc) => Monotype loc -> Monotype loc -> Eff r (Constraint loc, Substitution loc)

bindGiven :: forall (r :: [Effect]) loc. (StructuredDebug :> r, Error (UnifyError loc) :> r, Show loc) => TypeVariable -> Monotype loc -> Eff r (Substitution loc)

unifyVar :: forall loc (r :: [Effect]). (StructuredDebug :> r, Error (UnifyError loc) :> r, Reader (Set UniqueTyVar) :> r, Show loc, Eq loc) => UniqueTyVar -> Monotype loc -> Eff r (Constraint loc, Substitution loc)

unifyMany :: forall (r :: [Effect]) loc. (HasCallStack, StructuredDebug :> r, Pretty loc, Reader (Set UniqueTyVar) :> r, Eq loc, Show loc, Semigroup loc, Monoid (Constraint loc), ?constraint :: Maybe (Constraint loc), Error (UnifyError loc) :> r) => [Monotype loc] -> [Monotype loc] -> Eff r (Constraint loc, Substitution loc)

data UnifyError loc

Constructors

OccursCheckFailed TypeVariable (Monotype loc)

ScalarMismatch (loc, Scalar) (loc, Scalar)

TypeConstructorMismatch (loc, Qualified TypeName) (loc, Qualified TypeName)

ArityMismatch

UnificationFailed (Maybe (Constraint loc)) (Monotype loc, Monotype loc)

UnifyMismatch

UnresolvedConstraint (Qualified VarName) (Constraint loc)

PatternConstructorArityMismatch DataCon Int Int loc