Elara.TypeInfer.ConstraintGeneration

Documentation

generateConstraints :: forall (r :: [Effect]) loc. ConstraintGenEffects r loc => ShuntedExpr -> Eff r (TypedExpr, Monotype loc)

solveConstraint :: forall (r :: [Effect]) a. (UnifyEffects r a, Monoid (Constraint a)) => Constraint a -> Set UniqueTyVar -> Constraint a -> Eff r (Constraint a, Substitution a)

lookupType :: forall (r :: [Effect]) loc. (StructuredDebug :> r, QueryEffects r, loc ~ SourceRegion, Rock Query :> r, State (TypeEnvironment loc) :> r) => TypeEnvKey loc -> Eff r (Type loc)

Lookup a type in the type environment, querying if necessary. Looks in the local environment first to avoid unnecessary queries / infinite loops, then queries the global environment if not found locally.

unify :: forall loc (r :: [Effect]). (UnifyEffects r loc, Monoid (Constraint loc), ?constraint :: Maybe (Constraint loc), Reader (Set UniqueTyVar) :> r) => Monotype loc -> Monotype loc -> Eff r (Constraint loc, Substitution loc)