Rock.Memo

Implicit memoisation-

class HasIOE (f :: [Effect] -> Type -> Type) where

Proof that every key permits IO

Methods

withIOE :: forall (es :: [Effect]) a. f es a -> (IOE :> es => Eff es a) -> Eff es a

class HasMemoiseE (f :: [Effect] -> Type -> Type) where

Methods

withMemoiseE :: forall (es :: [Effect]) a. f es a -> ((Memoise :> es, Concurrent :> es) => Eff es a) -> Eff es a

memoise :: forall (f :: [Effect] -> Type -> Type). (forall (es :: [Effect]). GEq (f es), forall (es :: [Effect]) a. Hashable (f es a), HasMemoiseE f, HasCallStack) => Rules f -> Rules f

Remember what f queries have already been performed and their results in a DHashMap, and reuse them if a query is performed again a second time.

The DHashMap should typically not be reused if there has been some change that might make a query return a different result.

Explicit memoisation

data MemoQuery (f :: [(Type -> Type) -> Type -> Type] -> k -> Type) (es :: [(Type -> Type) -> Type -> Type]) (a :: k) where

Constructors

MemoQuery :: forall {k} (f :: [(Type -> Type) -> Type -> Type] -> k -> Type) (es1 :: [(Type -> Type) -> Type -> Type]) (a :: k). f es1 a -> MemoQuery f (IOE ': es1) a

withoutMemoisation :: forall (f :: [Effect] -> Type -> Type). Rules f -> Rules (MemoQuery f)

memoiseExplicit :: forall (f :: [Effect] -> Type -> Type). (forall (es :: [Effect]). GEq (f es), forall (es :: [Effect]) a. Hashable (f es a)) => IORef (DHashMap (HideEffects f) MVar) -> Rules f -> Rules (MemoQuery f)

Remember what f queries have already been performed and their results in a DHashMap, and reuse them if a query is performed again a second time.

The DHashMap should typically not be reused if there has been some change that might make a query return a different result.

newtype Cyclic (f :: k -> Type)

Constructors

Cyclic (Some f)

data MemoEntry a

Constructors

Started !ThreadId !(MVar (Maybe a)) !(MVar (Maybe [ThreadId]))

Done !a

memoiseWithCycleDetection :: forall (f :: [Effect] -> Type -> Type). (Typeable f, forall (es :: [Effect]) a. Show (f es a), forall (es :: [Effect]). GEq (f es), forall (es :: [Effect]) a. Hashable (f es a)) => IORef (DHashMap (HideEffects f) MemoEntry) -> IORef (HashMap ThreadId ThreadId) -> Rules f -> Rules (MemoQuery f)

Like memoise, but throw Cyclic f if a query depends on itself, directly or indirectly.

The HashMap represents dependencies between threads and should not be reused between invocations.